### 2024-09-16 #7229 @Kevin-Mattheus-Moerman
@hdlugas thanks for your submission to JOSS. To help check if this work is in scope for JOSS, I've just flagged it for scope review by our editorial board. This is because of the relatively small size of the project. Among other things, the scope check will determine if the project conforms to our [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion.
The scope review should take about 2 weeks to complete.
