### 2024-11-01 #7229 @Kevin-Mattheus-Moerman
@hdlugas the scope review has determined this work may be in scope for JOSS. I'll now look for a handling editor.
