### 2024-12-02 #7229 @jromanowska
Hi, @hdlugas :wave:
I will take care of the process from now on. I will try to find reviewers, but please let me know if you have any suggestions (or oppositions). When naming reviewers, write their GitHub handles _without_ `@`, so that they won't get notified.

When it comes to the paper, please reduce the word-count. to around 1000. One thing you can do is to remove the very long command examples - these don't really suit in JOSS paper, but rather in the documentation.
The last thing is to check the DOIs - as noted by editorialbot [above](https://github.com/openjournals/joss-reviews/issues/7229#issuecomment-2352762529), there are some references that have missing DOIs.
