+++ paper/paper.md
@@ -258,141 +258,9 @@ These tasks are implemented separately for (i) GC-MS and (ii) LC-MS data
 due to the different spectrum preprocessing transformations. To see all
 parameters for any of the five main scripts (build_library.py,
 spec_lib_matching_lcms.py, spec_lib_matching_gcms.py,
-plot_spectra_lcms.py, plot_spectra_gcms.py), run:
+plot_spectra_lcms.py, plot_spectra_gcms.py), see the documentation at 
+<a href="[url](https://github.com/hdlugas/MZsearch)">[link text](https://github.com/hdlugas/MZsearch)</a>.
 
-```         
-python build_library.py -h
-python spec_lib_matching_lcms.py -h
-python spec_lib_matching_lcms.py -h
-python plot_spectra_lcms.py -h
-python plot_spectra_gcms.py -h
-```
-
-The user can use their own libraries, and by default, MZsearch provides
-GC-MS and LC-MS reference libraries that have been trimmed. The
-recommended GC-MS and LC-MS reference libraries are available at the
-Zenodo database [@Dlugas2024_database]. The GC-MS library was extracted
-from the National Institute of Standards and Technology WebBook, while
-the LC-MS library was obtained from Global Natural Products Social
-Molecular Networking.
-
-## Get query/reference library from mgf, mzML, or cdf file
-
-To obtain a query/reference library in the necessary format for spectral
-library matching from an mgf, mzML, or cdf file, one can use:
-
-```         
-python build_library.py \
-  --input_path path_to_input_mgf_or_mzML_or_cdf_file \
-  --output_path path_to_output_csv_file 
-```
-
-## Run spectral library matching
-
-To run spectral library matching on LC-MS/GC-MS data, one can use:
-
-```         
-python spec_lib_matching_lcms.py \
-  --query_data path_to_query_lcms_CSV_file \
-  --reference_data path_to_reference_lcms_CSV_file \
-  --similarity_measure cosine \
-  --spectrum_preprocessing_order FCNMWL \
-  --high_quality_reference_library False \
-  --mz_min 0\
-  --mz_max 999999999999\
-  --int_min 0\
-  --int_max 999999999999\
-  --window_size_centroiding 0.5 \
-  --window_size_matching 0.5 \
-  --noise_threshold 0 \
-  --wf_mz 0 \
-  --wf_intensity 1 \
-  --LET_threshold 0 \
-  --entropy_dimension 1.1 \
-  --normalization_method standard \
-  --n_top_matches_to_save 1 \
-  --print_id_results False \
-  --output_identification path_to_lcms_identification_results_CSV \
-  --output_similarity_scores path_to_CSV_of_all_lcms_similarity_scores
-
-python spec_lib_matching_gcms.py \
-  --query_data path_to_query_gcms_CSV_file \
-  --reference_data path_to_reference_gcms_CSV_file \
-  --similarity_measure cosine \
-  --spectrum_preprocessing_order FNLW \
-  --high_quality_reference_library False \
-  --mz_min 0\
-  --mz_max 999999999999\
-  --int_min 0\
-  --int_max 999999999999\
-  --wf_mz 0 \
-  --wf_intensity 1 \
-  --entropy_dimension 1.1 \
-  --normalization_method standard \
-  --n_top_matches_to_save 1 \
-  --print_id_results False \
-  --output_identification path_to_gcms_identification_results_CSV \
-  --output_similarity_scores path_to_CSV_of_all_gcms_similarity_scores
-```
-
-## Plot a query spectrum against a reference spectrum before and after spectrum preprocessing transformations
-
[…]
