### 2024-12-02 e314c1e
Update paper.bib
