+++ paper/paper.bib
@@ -6,7 +6,8 @@
   volume = {52},
   pages = {479--487},
   year = {1988},
-  publisher = {Springer}
+  publisher = {Springer},
+  doi = {10.1007/bf01016429}
 }
 
 @article{Huber2020,
@@ -45,7 +46,8 @@
   pages = {1158--1163},
   month = {apr},
   year = {2012},
-  language = {en}
+  language = {en},
+  doi = {10.1093/bioinformatics/bts083}
 }
 
 @Article{Li2021,
@@ -78,7 +80,8 @@
   author = {{Oh}, Y. and {Kim}, S. and {Jeong}, J},
   journal = {Chemometrics and intelligent laboratory systems : an international journal sponsored by the Chemometrics Society},
   title = {ShinyMetID: An R shiny package for metabolite identification by mass spectral matching.},
-  year = {2023}
+  year = {2023},
+  doi = {https://doi.org/10.1016/j.chemolab.2023.104861}
 }
 
 @article{Wu2021,
