### 2024-12-02 f836fe2
Update paper.md
