+++ README.md
@@ -4,7 +4,8 @@ Command-line Python tool to perform spectral library matching to identify chemic
 ## Table of Contents
 - [1. Create conda environment and install dependencies](#create-conda-env)
 - [2. Functionality](#functionality)
-- [2.1 Spectrum Preprocessing Transformations](#spec-preprocessing-transformations)
+   - [2.1 Spectrum Preprocessing Transformations](#spec-preprocessing-transformations)
+   - [2.2 Similarity Measures](#similarity-measures)
 - [3. Usage](#usage)
    - [3.1 Obtain LC-MS or GC-MS library from MGF, mzML, or cdf file](#process-data)
    - [3.2 Run spectral library matching](#run-spec-lib-matching)
@@ -38,6 +39,113 @@ transformations is offered in MZsearch:
     the transformed spectrum $I^{\star}$ consists of the peaks
     $(m_{i},x_{i})$ in $I$ such that mz_min $\leq m_{i}\leq$ mz_max and
     int_min $\leq x_{i}\leq$ int_max.
+    -   Weight Factor Transformation: Given a pair of user-defined weight
+    factor parameters $(\text{a,b})$ and spectrum $I$ with m/z values
+    $(m_{1},m_{2},...,m_{n})$ and intensities $(x_{1},x_{2},...,x_{n})$,
+    the transformed spectrum $I^{\star}$ has the same m/z values as $I$
+    and has intensities given by
+    $I^{\star}:=(m_{1}^{\text{a}}\cdot x_{1}^{\text{b}},m_{2}^{\text{a}}\cdot x_{2}^{\text{b}},...,m_{n}^{\text{a}}\cdot x_{n}^{\text{b}})$.
+
+-   Low-Entropy Transformation: Given a user-defined low-entropy
+    threshold parameter $T$ and spectrum $I$ with intensities
+    $(x_{1},x_{2},...,x_{n})$, $\sum_{i=1}^nx_i = 1$, and Shannon
+    entropy $H_{Shannon}(I)=-\sum_{i=1}^{n}x_{i}\cdot ln(x_{i})$, the
+    transformed spectrum intensities
+    $I^{\star}=(x_{1}^{\star},x_{2}^{\star},...,x_{n}^{\star})$ are such
+    that, for all $i\in\{1,2,...,n\}$, $x_{i}^{\star}=x_{i}$ if
+    $H_{Shannon}(I)\geq T$ and
+    $x_{i}^{\star}=x_{i}^{\frac{1+H_{Shannon}(I)}{1+T}}$ if
+    $H_{Shannon}(I)\textless T$.
+
+-   Centroiding (only applicable to LC-MS data): Given a user-defined
+    window-size parameter $w_{centroiding}$ and a spectrum $I$ with m/z
+    values $(m_{1},m_{2},...,m_{n})$ and intensities
+    $(x_{1},x_{2},...,x_{n})$, the transformed spectrum $I^{\star}$
+    merges adjacent peaks $(m_{i},x_{i}),(m_{i+1},x_{i+1})$ into the
+    peak
+    $(\frac{m_{i}\cdot x_{i}+m_{i+1}\cdot x_{i+1}}{x_{i}+x_{i+1}},x_{i}+x_{i+1})$
+    if $|m_{i}-m_{i+1}|\textless w_{centroiding}$ for
+    $i\in\{1,2,...,n-1\}$. This centroiding procedure generalizes to
+    more than two peaks whose m/z values are within distance
[…]
