+++ README.md
@@ -117,31 +117,27 @@ with $0\leq a_{i},b_{i}\leq 1$ for all $i\in\{1,2,...,n\}$ and
 $\sum_{i=1}^{n}a_{i}=\sum_{i=1}^{n}b_{i}=1$, MZsearch provides
 functionality for computing the following similarity measures:
 
--   Cosine Similarity Measure: $S_{Cosine}(I,J)=\frac{I\circ J}{|I|_{2}\cdot |J|_{2}}$ where multiplication in the numerator refers to the
+-   Cosine Similarity Measure: $S_{Cosine}(I,J)=\frac{I\circ J}{|I|_{2}\cdot |J|_{2}}$ where multiplication in the numerator refers to $\frac{1}{2}$ the
     dot product $I\circ J=a_{1}b_{1}+a_{2}b_{2}+...+a_{n}b_{n}$ of $I$
     and $J$ and multiplication in the denominator refers to
     multiplication of the $L^{2}$-norm of $I$ and $J$,
     $|I|_{2}=\sqrt{a_{1}^{2}+a_{2}^{2}+...+a_{n}^{2}}, |J|_{2}=\sqrt{b_{1}^{2}+b_{2}^{2}+...+b_{n}^{2}}$.
 
--   Shannon Entropy Similarity Measure: \begin{gather*}
+-   Shannon Entropy Similarity Measure:
       S_{Shannon}(I,J) = 1-\frac{2\cdot H_{Shannon}\left(\frac{I+J}{2}\right) - H_{Shannon}(I)-H_{Shannon}(J)}{ln(4)},\\
       H_{Shannon}(I)=-\sum_{i=1}^{n}a_{i}\cdot ln(a_{i})
-    \end{gather*}
 
 -    Tsallis Entropy Similarity Measure:
-    \begin{gather*}\label{eq:}
       S_{}(I,J,q)=1-\frac{2\times H_{}(I/2+J/2,q)-H_{}(I,q)-H_{}(J,q)}{N_{}(I,J,q)},\\
       N_{}(I,J,q):=\frac{\sum_{i=1}^{n}\left(2\left(\frac{a_{i}}{2}\right)^{q}+2\left(\frac{b_{i}}{2}\right)^{q}-a_{i}^{q}-b_{i}^{q}\right)}{1-q},\\
       H_{}(I,q)=\frac{\left(\sum_{i=1}^{n}a_{i}^{q}\right)-1}{1-q},\\
       q\neq 1, \ q\textgreater 0
-    \end{gather*}
 
--   Rényi Entropy Similarity Measure: \begin{gather*}\label{eq:renyi}
+-   Rényi Entropy Similarity Measure:
       S_{Renyi}(I,J,q)=1-\frac{2\times H_{Renyi}(I/2+J/2,q)-H_{Renyi}(I,q)-H_{Renyi}(J,q)}{N_{Renyi}(I,J,q)},\\
       N_{Renyi}(I,J,q):=\left(\frac{1}{1-q}\right)\left(2\times ln\left(\sum_{i}(a_{i}/2)^{q}+\sum_{j}(b_{j}/2)^{q}\right)-ln(\sum_{i}a_{i}^{q})-ln(\sum_{i}b_{i}^{q})\right),\\
       H_{Renyi}(I,q)=\frac{1}{1-q}ln(\sum_{i=1}^{n}a_{i}^{q}),\\
       q\neq 1, \ q\textgreater 0
-    \end{gather*}
 
 
 
