+++ paper/paper.md
@@ -124,84 +124,21 @@ recently-developed  Entropy Similarity Measure
 ## Spectrum Preprocessing Transformations
 
 Functionality implementing the following spectrum preprocessing
-transformations is offered in MZsearch:
+transformations is offered in MZsearch (more details found in documentation: <a href="[url](https://github.com/hdlugas/MZsearch)"></a>):
 
--   Filtering: Given user-defined parameters (mz_min,mz_max),
-    (int_min,int_max) and spectrum $I$ with m/z values
-    $(m_{1},m_{2},...,m_{n})$ and intensities $(x_{1},x_{2},...,x_{n})$,
-    the transformed spectrum $I^{\star}$ consists of the peaks
-    $(m_{i},x_{i})$ in $I$ such that mz_min $\leq m_{i}\leq$ mz_max and
-    int_min $\leq x_{i}\leq$ int_max.
+-   Filtering based on mz-values and intensities.
 
--   Weight Factor Transformation: Given a pair of user-defined weight
-    factor parameters $(\text{a,b})$ and spectrum $I$ with m/z values
-    $(m_{1},m_{2},...,m_{n})$ and intensities $(x_{1},x_{2},...,x_{n})$,
-    the transformed spectrum $I^{\star}$ has the same m/z values as $I$
-    and has intensities given by
-    $I^{\star}:=(m_{1}^{\text{a}}\cdot x_{1}^{\text{b}},m_{2}^{\text{a}}\cdot x_{2}^{\text{b}},...,m_{n}^{\text{a}}\cdot x_{n}^{\text{b}})$.
+-   Weight Factor Transformation.
 
--   Low-Entropy Transformation: Given a user-defined low-entropy
-    threshold parameter $T$ and spectrum $I$ with intensities
-    $(x_{1},x_{2},...,x_{n})$, $\sum_{i=1}^nx_i = 1$, and Shannon
-    entropy $H_{Shannon}(I)=-\sum_{i=1}^{n}x_{i}\cdot ln(x_{i})$, the
-    transformed spectrum intensities
-    $I^{\star}=(x_{1}^{\star},x_{2}^{\star},...,x_{n}^{\star})$ are such
-    that, for all $i\in\{1,2,...,n\}$, $x_{i}^{\star}=x_{i}$ if
-    $H_{Shannon}(I)\geq T$ and
-    $x_{i}^{\star}=x_{i}^{\frac{1+H_{Shannon}(I)}{1+T}}$ if
-    $H_{Shannon}(I)\textless T$.
+-   Low-Entropy Transformation.
 
--   Centroiding (only applicable to LC-MS data): Given a user-defined
-    window-size parameter $w_{centroiding}$ and a spectrum $I$ with m/z
-    values $(m_{1},m_{2},...,m_{n})$ and intensities
-    $(x_{1},x_{2},...,x_{n})$, the transformed spectrum $I^{\star}$
-    merges adjacent peaks $(m_{i},x_{i}),(m_{i+1},x_{i+1})$ into the
-    peak
-    $(\frac{m_{i}\cdot x_{i}+m_{i+1}\cdot x_{i+1}}{x_{i}+x_{i+1}},x_{i}+x_{i+1})$
-    if $|m_{i}-m_{i+1}|\textless w_{centroiding}$ for
-    $i\in\{1,2,...,n-1\}$. This centroiding procedure generalizes to
-    more than two peaks whose m/z values are within distance
-    $w_{centroiding}$ of each other.
+-   Centroiding (only applicable to LC-MS data).
 
--   Noise Removal: Given a user-defined noise removal parameter $r$ and
[…]
