### 2024-12-02 ad0a784
Update paper.md
