+++ paper/paper.bib
@@ -72,7 +72,7 @@
   year = {1967},
   volume = {3},
   pages = {30-35},
-  url = {https://api.semanticscholar.org/CorpusID:14939337}
+  url = {http://dml.cz/dmlcz/125526}
 }
 
 @article{Kim2023,
