### 2024-12-16 #7229 @hdlugas
@jromanowska Yep, we think the following may have a background suitable for reviewers of our software (GitHub usernames are in parentheses):
* Jianhui Gong (duncanjunegong)
* Lorenzo Castellino (Kastakin)
* Senthil Murugapiran (aboffin)
* adafede (adafede)
* askerdb (askerdb)
* florian-huber (florian-huber).

Lorenzo Castellino and Senthil Murugapiran may be particularly well-suited to review since (i) our software is a Python-based tool to be used in the Bash shell and (ii) these two potential reviewers are familiar with Bash.

Note that we selected possible reviewers as those with an area of expertise in "Biomedical Engineering, Biosciences, Chemistry, and Materials" with Python experience and keyword "metabolomics".
