### 2024-12-17 #7594 @editorialbot
[REVIEW]: MZsearch: A Python-Based Compound Identification Tool for GC-MS and LC-MS-Based Metabolomics
**Submitting author:** @hdlugas (<a href="http://orcid.org/0000-0002-6819-0045">Hunter Dlugas</a>)
**Repository:** https://github.com/hdlugas/MZsearch
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
