### 2024-12-17 e630e5e
Update paper.bib
removed redundant "https://doi.org/" from doi part of reference
