+++ paper/paper.bib
@@ -61,8 +61,7 @@
   number = {12},
   pages = {1524-1531},
   issn = {1548-7105},
-  doi = {10.1038/s41592-021-01331-z},
-  url = {https://doi.org/10.1038/s41592-021-01331-z}
+  doi = {10.1038/s41592-021-01331-z}
 }
 
 @article{Havrda1967,
@@ -81,7 +80,7 @@
   journal = {Chemometrics and intelligent laboratory systems : an international journal sponsored by the Chemometrics Society},
   title = {ShinyMetID: An R shiny package for metabolite identification by mass spectral matching.},
   year = {2023},
-  doi = {https://doi.org/10.1016/j.chemolab.2023.104861}
+  doi = {10.1016/j.chemolab.2023.104861}
 }
 
 @article{Wu2021,
@@ -95,7 +94,6 @@
   month = {08},
   issn = {1367-4803},
   doi = {10.1093/bioinformatics/btab583},
-  url = {https://doi.org/10.1093/bioinformatics/btab583},
   eprint = {https://academic.oup.com/bioinformatics/article-pdf/38/2/568/49006586/btab583.pdf},
 }
 
