### 2025-01-08 #7594 @aboffin
@jromanowska Thank you, hope you had a good break also. I have cloned the repo and reading the code, related documents, I need another 2-3 weeks to complete the review. Hope this is acceptable.

In addition, I do not understand these two items in the checklist:

Reproducibility: If the paper contains original results, results are entirely reproducible by reviewers. If the paper contains no original results, please check this item.

Human and animal research: If the paper contains original data research on humans subjects or animals, does it comply with [JOSS's human participants research policy and/or animal research policy](https://joss.readthedocs.io/en/latest/policies.html?highlight=animal#joss-policies)? If the paper contains no such data, please check this item.

For the above two items it explicitly requires the reviewer to check the items (or not), whereas one would assume that they would go through the entire checklist and tick items that are applicable? Thank you in advance for the clarification.
