### 2025-01-09 #7594 @jromanowska
Thank you for the update, @aboffin. No problem with waiting 2-3 weeks.

Apropos the two items in the checklist you've mentioned: it only says that you should check the item anyway _if the paper does not present results_ or _if the paper does not contain animal or human research data_. If the paper does contain those, you should of course check the item _only_ if it complies with the requirements.
