+++ tests/tests_plot_spectra_gcms.sh
@@ -1,29 +1,25 @@
 #!/bin/bash
-
-source activate base
-conda activate mzsearch_env
-
-cd /home/hunter/mass_spec/entropies/JOSS/scripts
-
+# conda activate mzsearch_env
+cd ../src
 
 echo $'\n\n\n\n\ntest #0'
 python plot_spectra_gcms.py \
-  --query_data /home/hunter/mass_spec/entropies/JOSS/data/gcms_query_library.csv \
-  --reference_data /home/hunter/mass_spec/entropies/JOSS/data/gcms_reference_library.csv \
+  --query_data ../data/gcms_query_library.csv \
+  --reference_data ../data/gcms_reference_library.csv \
   --similarity_measure cosine \
   --normalization_method standard \
-  --save_plots /home/hunter/mass_spec/entropies/JOSS/scripts/test1.pdf
+  --save_plots ../src/test1.pdf
 
 echo $'\n\n\n\n\ntest #1'
 python plot_spectra_gcms.py \
-  --query_data /home/hunter/mass_spec/entropies/JOSS/data/gcms_query_library.csv \
-  --reference_data /home/hunter/mass_spec/entropies/JOSS/data/gcms_reference_library.csv \
+  --query_data ../data/gcms_query_library.csv \
+  --reference_data ../data/gcms_reference_library.csv \
   --similarity_measure tsallis \
   --wf_mz 2 \
   --wf_intensity 0.5 \
   --normalization_method standard \
   --entropy_dimension 2 \
-  --save_plots /home/hunter/mass_spec/entropies/JOSS/scripts/test1.pdf
+  --save_plots ../src/test1.pdf
 
 echo $'\n\n\n\n\ntest #2'
 python plot_spectra_gcms.py \
@@ -32,46 +28,46 @@ python plot_spectra_gcms.py \
   --wf_intensity 0.5 \
   --normalization_method standard \
   --entropy_dimension 0.5 \
-  --save_plots /home/hunter/mass_spec/entropies/JOSS/scripts/test2.pdf
+  --save_plots ../src/test2.pdf
 
 echo $'\n\n\n\n\ntest #3'
 python plot_spectra_gcms.py \
   --similarity_measure tsallis \
   --normalization_method standard \
   --entropy_dimension 2 \
-  --save_plots /home/hunter/mass_spec/entropies/JOSS/scripts/test3.pdf
+  --save_plots ../src/test3.pdf
 
 echo $'\n\n\n\n\ntest #4'
 python plot_spectra_gcms.py \
   --similarity_measure renyi \
   --normalization_method standard \
   --entropy_dimension 1.1 \
-  --save_plots /home/hunter/mass_spec/entropies/JOSS/scripts/test4.pdf
+  --save_plots ../src/test4.pdf
 
 echo $'\n\n\n\n\ntest #5'
 python plot_spectra_gcms.py \
   --similarity_measure renyi \
   --normalization_method standard \
   --entropy_dimension 0.9 \
-  --save_plots /home/hunter/mass_spec/entropies/JOSS/scripts/test5.pdf
+  --save_plots ../src/test5.pdf
 
 echo $'\n\n\n\n\ntest #6'
 python plot_spectra_gcms.py \
   --similarity_measure shannon \
   --normalization_method standard \
-  --save_plots /home/hunter/mass_spec/entropies/JOSS/scripts/test6.pdf
+  --save_plots ../src/test6.pdf
 
 echo $'\n\n\n\n\ntest #7'
 python plot_spectra_gcms.py \
   --similarity_measure cosine \
[…]
