### 2025-02-07 #7594 @aboffin
@jromanowska Please see the pull request with a few changes that I was able to make this week: https://github.com/hdlugas/MZsearch/pull/1

I will further work on the review while the author makes the requested changes.
