### 2025-02-10 #7594 @aboffin
@jromanowska It was FYI only, sorry for not being clear. The author @hdlugas was not one of the options when I tried to include them, but I see now that it does show up if you type in their username. Thank you.
