+++ README.md
@@ -13,17 +13,8 @@ Command-line Python tool to perform spectral library matching to identify chemic
 - [4. Bugs/Questions?](#bugs-questions)
 
 <a name="create-conda-env"></a>
-## 1. Create conda environment and install dependencies
-MZsearch requires the Python dependencies Matplotlib, NumPy, Pandas, and SciPy, Pyteomics, and netCDF4. Specifically, this software was validated with python=3.12.4, matplotlib=3.8.4, numpy=1.26.4, pandas=2.2.2, scipy=1.13.1, pyteomics=4.7.2, and nedCDF4=1.6.5, although it may work with other versions of these tools. For instructions on installing conda on your system, see: [https://docs.conda.io/projects/conda/en/latest/user-guide/install/index.html](https://docs.conda.io/projects/conda/en/latest/user-guide/install/index.html). Once conda is installed, clone the GitHub repository, navigate to the directory containing the environment.yml file, and create (and activate) a conda environment with the necessary dependencies with the commands:
-```
-conda env create -f environment.yml
-conda activate mzsearch_env
-```
-
-To return to your base environment, you can deactivate the mzsearch_env environment with the command:
-```
-conda deactivate
-```
+## 1. Install dependencies
+MZsearch requires the Python dependencies Matplotlib, NumPy, Pandas, and SciPy, Pyteomics, and netCDF4. Specifically, this software was validated with python=3.12.4, matplotlib=3.8.4, numpy=1.26.4, pandas=2.2.2, scipy=1.13.1, pyteomics=4.7.2, and nedCDF4=1.6.5, although it may work with other versions of these tools. A user may consider creating a conda environment (see [https://docs.conda.io/projects/conda/en/latest/user-guide/install/index.html](https://docs.conda.io/projects/conda/en/latest/user-guide/install/index.html)).
 
 <a name="functionality"></a>
 ## 2. Functionality
