### 2025-02-11 f1a1bd9
Delete tests/tests_plot_spectra_gcms.sh
