### 2025-02-11 3d24ceb
Delete tests/tests_plot_spectra_lcms.sh
