+++ tests/tests_spec_lib_matching_gcms.sh
@@ -1,161 +1,0 @@
-#!/bin/bash
-
-cd ../src
-
-echo $'\n\n\n\n\ntest #0'
-python spec_lib_matching_gcms.py \
-  --query_data ../data/gcms_query_library.csv \
-  --reference_data ../data/gcms_reference_library.csv \
-  --similarity_measure cosine \
-  --normalization_method standard \
-  --print_id_results True \
-
-echo $'\n\n\n\n\ntest #1'
-python spec_lib_matching_gcms.py \
-  --query_data ../data/gcms_query_library.csv \
-  --reference_data ../data/gcms_reference_library.csv \
-  --similarity_measure tsallis \
-  --wf_mz 2 \
-  --wf_intensity 0.5 \
-  --normalization_method standard \
-  --entropy_dimension 2 \
-  --n_top_matches_to_save 4 \
-  --print_id_results True \
-  --output_identification ../output_gcms_identification.csv \
-  --output_similarity_scores ../output_gcms_all_similarity_scores.csv
-
-echo $'\n\n\n\n\ntest #2'
-python spec_lib_matching_gcms.py \
-  --similarity_measure tsallis \
-  --wf_mz 2 \
-  --wf_intensity 0.5 \
-  --normalization_method standard \
-  --entropy_dimension 0.5 \
-  --n_top_matches_to_save 2 \
-  --print_id_results True \
-
-echo $'\n\n\n\n\ntest #3'
-python spec_lib_matching_gcms.py \
-  --similarity_measure tsallis \
-  --normalization_method standard \
-  --entropy_dimension 2 \
-  --print_id_results True
-
-echo $'\n\n\n\n\ntest #4'
-python spec_lib_matching_gcms.py \
-  --similarity_measure renyi \
-  --normalization_method standard \
-  --entropy_dimension 1.1 \
-  --n_top_matches_to_save 2 \
-  --print_id_results True
-
-echo $'\n\n\n\n\ntest #5'
-python spec_lib_matching_gcms.py \
-  --similarity_measure renyi \
-  --normalization_method standard \
-  --entropy_dimension 0.9 \
-  --n_top_matches_to_save 2 \
-  --print_id_results True
-
-echo $'\n\n\n\n\ntest #6'
-python spec_lib_matching_gcms.py \
-  --similarity_measure shannon \
-  --normalization_method standard \
-  --n_top_matches_to_save 2 \
-  --print_id_results True
-
-echo $'\n\n\n\n\ntest #7'
-python spec_lib_matching_gcms.py \
-  --similarity_measure cosine \
-  --n_top_matches_to_save 2 \
-  --print_id_results True
-
-echo $'\n\n\n\n\ntest #8'
-python spec_lib_matching_gcms.py \
-  --spectrum_preprocessing_order FNLW\
-  --similarity_measure cosine \
-  --wf_mz 0.5 \
-  --n_top_matches_to_save 2 \
-  --print_id_results True
-
-echo $'\n\n\n\n\ntest #9'
-python spec_lib_matching_gcms.py \
-  --spectrum_preprocessing_order FNLW\
-  --similarity_measure cosine \
-  --mz_min 200 \
-  --mz_max 250 \
-  --int_min 50 \
-  --int_max 500 \
-  --n_top_matches_to_save 2 \
-  --print_id_results True
-
-echo $'\n\n\n\n\ntest #10'
-python spec_lib_matching_gcms.py \
-  --spectrum_preprocessing_order FNLW\
-  --similarity_measure cosine \
-  --mz_max 100 \
-  --n_top_matches_to_save 2 \
-  --print_id_results True
-
[…]
