### 2025-02-11 74f2a8f
Delete tests/tests_spec_lib_matching_lcms.sh
