+++ tests/tests_spec_lib_matching_lcms.sh
@@ -1,254 +1,0 @@
-#!/bin/bash
-
-cd ../src
-
-echo $'\n\n\n\n\ntest #1'
-python spec_lib_matching_lcms.py \
-  --query_data ../data/lcms_query_library.csv \
-  --reference_data ../data/lcms_reference_library.csv \
-  --similarity_measure tsallis \
-  --wf_mz 2 \
-  --wf_intensity 0.5 \
-  --normalization_method standard \
-  --entropy_dimension 2 \
-  --n_top_matches_to_save 4 \
-  --print_id_results True \
-  --output_identification ../output_lcms_identification.csv \
-  --output_similarity_scores ../output_lcms_all_similarity_scores.csv
-
-echo $'\n\n\n\n\ntest #2'
-python spec_lib_matching_lcms.py \
-  --similarity_measure tsallis \
-  --wf_mz 2 \
-  --wf_intensity 0.5 \
-  --normalization_method standard \
-  --entropy_dimension 0.5 \
-  --n_top_matches_to_save 2 \
-  --print_id_results True \
-
-echo $'\n\n\n\n\ntest #3'
-python spec_lib_matching_lcms.py \
-  --similarity_measure tsallis \
-  --normalization_method standard \
-  --entropy_dimension 2 \
-  --print_id_results True
-
-echo $'\n\n\n\n\ntest #4'
-python spec_lib_matching_lcms.py \
-  --similarity_measure renyi \
-  --normalization_method standard \
-  --entropy_dimension 1.1 \
-  --n_top_matches_to_save 2 \
-  --print_id_results True
-
-
-echo $'\n\n\n\n\ntest #5'
-python spec_lib_matching_lcms.py \
-  --similarity_measure renyi \
-  --normalization_method standard \
-  --entropy_dimension 0.9 \
-  --n_top_matches_to_save 2 \
-  --print_id_results True
-
-echo $'\n\n\n\n\ntest #6'
-python spec_lib_matching_lcms.py \
-  --similarity_measure shannon \
-  --normalization_method standard \
-  --n_top_matches_to_save 2 \
-  --print_id_results True
-
-echo $'\n\n\n\n\ntest #7'
-python spec_lib_matching_lcms.py \
-  --similarity_measure cosine \
-  --n_top_matches_to_save 2 \
-  --print_id_results True
-
-echo $'\n\n\n\n\ntest #8'
-python spec_lib_matching_lcms.py \
-  --spectrum_preprocessing_order LFWNCM \
-  --similarity_measure cosine \
-  --wf_mz 0.5 \
-  --n_top_matches_to_save 2 \
-  --print_id_results True
-
-echo $'\n\n\n\n\ntest #9'
-python spec_lib_matching_lcms.py \
-  --spectrum_preprocessing_order LFWNCM \
-  --similarity_measure cosine \
-  --mz_min 200 \
-  --mz_max 250 \
-  --int_min 50 \
-  --int_max 500 \
-  --n_top_matches_to_save 2 \
-  --print_id_results True
-
-echo $'\n\n\n\n\ntest #10'
-python spec_lib_matching_lcms.py \
-  --spectrum_preprocessing_order FWCNLM \
-  --similarity_measure cosine \
-  --mz_max 300 \
-  --n_top_matches_to_save 2 \
-  --print_id_results True
-
-echo $'\n\n\n\n\ntest #11'
-python spec_lib_matching_lcms.py \
-  --spectrum_preprocessing_order FWCNLM \
-  --similarity_measure cosine \
-  --int_max 400 \
[…]
