+++ tests/tests_spec_lib_matching.sh
@@ -1,0 +1,542 @@
+#!/bin/bash
+
+cd /home/hunter/mass_spec/entropies/JOSS/scripts/reviewer1
+
+
+echo $'\n\n\n\n\ntest #1'
+python spec_lib_matching.py \
+  --query_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_query_library.csv \
+  --reference_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_reference_library.csv \
+  --chromatography_platform LCMS \
+  --similarity_measure tsallis \
+  --wf_mz 2 \
+  --wf_intensity 0.5 \
+  --normalization_method standard \
+  --entropy_dimension 2 \
+  --n_top_matches_to_save 4 \
+  --print_id_results True \
+  --output_identification /home/hunter/mass_spec/entropies/JOSS/scripts/reviewer1/output_lcms_identification.csv \
+  --output_similarity_scores /home/hunter/mass_spec/entropies/JOSS/scripts/reviewer1/output_lcms_all_similarity_scores.csv
+
+echo $'\n\n\n\n\ntest #2'
+python spec_lib_matching.py \
+  --query_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_query_library.csv \
+  --reference_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_reference_library.csv \
+  --chromatography_platform LCMS \
+  --similarity_measure tsallis \
+  --wf_mz 2 \
+  --wf_intensity 0.5 \
+  --normalization_method standard \
+  --entropy_dimension 0.5 \
+  --n_top_matches_to_save 2 \
+  --print_id_results True \
+
+echo $'\n\n\n\n\ntest #3'
+python spec_lib_matching.py \
+  --query_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_query_library.csv \
+  --reference_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_reference_library.csv \
+  --chromatography_platform LCMS \
+  --similarity_measure tsallis \
+  --normalization_method standard \
+  --entropy_dimension 2 \
+  --print_id_results True
+
+echo $'\n\n\n\n\ntest #4'
+python spec_lib_matching.py \
+  --query_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_query_library.csv \
+  --reference_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_reference_library.csv \
+  --chromatography_platform LCMS \
+  --similarity_measure renyi \
+  --normalization_method standard \
+  --entropy_dimension 1.1 \
+  --n_top_matches_to_save 2 \
+  --print_id_results True
+
+
+echo $'\n\n\n\n\ntest #5'
+python spec_lib_matching.py \
+  --query_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_query_library.csv \
+  --reference_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_reference_library.csv \
+  --chromatography_platform LCMS \
+  --similarity_measure renyi \
+  --normalization_method standard \
+  --entropy_dimension 0.9 \
+  --n_top_matches_to_save 2 \
+  --print_id_results True
+
+echo $'\n\n\n\n\ntest #6'
+python spec_lib_matching.py \
+  --query_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_query_library.csv \
+  --reference_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_reference_library.csv \
+  --chromatography_platform LCMS \
+  --similarity_measure shannon \
[…]
