+++ tests/tests_plot_spectra.sh
@@ -1,0 +1,565 @@
+#!/bin/bash
+
+cd /home/hunter/mass_spec/entropies/JOSS/scripts/reviewer1
+
+
+echo $'\n\n\n\n\ntest #1'
+python plot_spectra.py \
+  --query_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_query_library.csv \
+  --reference_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_reference_library.csv \
+  --similarity_measure tsallis \
+  --chromatography_platform LCMS \
+  --wf_mz 2 \
+  --wf_intensity 0.5 \
+  --normalization_method standard \
+  --entropy_dimension 2 \
+  --save_plots /home/hunter/mass_spec/entropies/JOSS/scripts/reviewer1/test1.pdf
+
+echo $'\n\n\n\n\ntest #2'
+python plot_spectra.py \
+  --query_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_query_library.csv \
+  --reference_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_reference_library.csv \
+  --similarity_measure tsallis \
+  --chromatography_platform LCMS \
+  --wf_mz 2 \
+  --wf_intensity 0.5 \
+  --normalization_method standard \
+  --entropy_dimension 2 \
+  --save_plots /home/hunter/mass_spec/entropies/JOSS/scripts/reviewer1/test2.pdf
+
+echo $'\n\n\n\n\ntest #2'
+python plot_spectra.py \
+  --query_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_query_library.csv \
+  --reference_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_reference_library.csv \
+  --similarity_measure tsallis \
+  --chromatography_platform LCMS \
+  --wf_mz 2 \
+  --wf_intensity 0.5 \
+  --normalization_method standard \
+  --entropy_dimension 0.5 \
+  --save_plots /home/hunter/mass_spec/entropies/JOSS/scripts/reviewer1/test2.pdf
+
+echo $'\n\n\n\n\ntest #3'
+python plot_spectra.py \
+  --query_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_query_library.csv \
+  --reference_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_reference_library.csv \
+  --similarity_measure tsallis \
+  --chromatography_platform LCMS \
+  --normalization_method standard \
+  --entropy_dimension 2 \
+  --save_plots /home/hunter/mass_spec/entropies/JOSS/scripts/reviewer1/test3.pdf
+
+echo $'\n\n\n\n\ntest #4'
+python plot_spectra.py \
+  --query_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_query_library.csv \
+  --reference_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_reference_library.csv \
+  --similarity_measure renyi \
+  --chromatography_platform LCMS \
+  --normalization_method standard \
+  --entropy_dimension 1.1 \
+  --save_plots /home/hunter/mass_spec/entropies/JOSS/scripts/reviewer1/test4.pdf
+
+
+echo $'\n\n\n\n\ntest #5'
+python plot_spectra.py \
+  --query_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_query_library.csv \
+  --reference_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_reference_library.csv \
+  --similarity_measure renyi \
+  --chromatography_platform LCMS \
[…]
