### 2025-02-11 9b20b8a
Delete src/plot_spectra_gcms.py
