### 2025-02-11 6a47804
Delete src/plot_spectra_lcms.py
