+++ README.md
@@ -167,13 +167,11 @@ This repository has three main capabilities:
 2. running spectral library matching to identify compounds based off of their mass spectrometry data
 3. plotting a query spectrum vs a reference spectrum before and after preprocessing transformations.
 
-These tasks are implemented separately for the cases of (i) GC-MS and (ii) LC-MS data due to the different spectrum preprocessing transformations stemming from a different format in the mass to charge ratios in GC-MS vs LC-MS data. To see all parameters for any of the four main scripts (spec_lib_matching_lcms.py, spec_lib_matching_gcms.py, plot_spectra_lcms.py, plot_spectra_gcms.py), run:
+These tasks are implemented separately for the cases of (i) GC-MS and (ii) LC-MS data due to the different spectrum preprocessing transformations stemming from a different format in the mass to charge ratios in GC-MS vs LC-MS data. To see all parameters for any of the two main scripts (spec_lib_matching.py, plot_spectra.py), run:
 ```
 python build_library.py -h
-python spec_lib_matching_.py -h
-python spec_lib_matching_.py -h
-python plot_spectra_.py -h
-python plot_spectra_gcms.py -h
+python spec_lib_matching.py -h
+python plot_spectra.py -h
 ```
 
 <a name="process-data"></a>
@@ -199,18 +197,16 @@ To run spectral library matching on LC-MS/GC-MS data, one can use:
 ```
 python spec_lib_matching_.py \
   --query_data path_to_query__CSV_file \
-  --reference_data path_to_reference__CSV_file
-
-python spec_lib_matching_gcms.py \
-  --query_data path_to_query_gcms_CSV_file \
-  --reference_data path_to_reference_gcms_CSV_file
+  --reference_data path_to_reference__CSV_file \
+  --chromatography_platform insert_platform
 ```
 
 Example implementations of these scripts with all parameters specified are:
 ```
-python spec_lib_matching_.py \
+python spec_lib_matching.py \
   --query_data path_to_query__CSV_file \
   --reference_data path_to_reference__CSV_file \
+  --chromatography_platform LCMS \
   --likely_reference_IDs None \
   --similarity_measure cosine \
   --spectrum_preprocessing_order FCNMWL \
@@ -231,37 +227,19 @@ python spec_lib_matching_.py \
   --print_id_results False \
   --output_identification path_to__identification_results_CSV \
   --output_similarity_scores path_to_CSV_of_all__similarity_scores
-
-python spec_lib_matching_gcms.py \
-  --query_data path_to_query_gcms_CSV_file \
-  --reference_data path_to_reference_gcms_CSV_file \
-  --likely_reference_IDs None \
-  --similarity_measure cosine \
-  --spectrum_preprocessing_order FNLW \
-  --high_quality_reference_library False \
-  --mz_min 0 \
-  --mz_max 999999999999 \
-  --int_min 0 \
-  --int_max 999999999999 \
-  --wf_mz 0 \
-  --wf_intensity 1 \
-  --entropy_dimension 1.1 \
-  --normalization_method standard \
-  --n_top_matches_to_save 1 \
-  --print_id_results False \
-  --output_identification path_to_gcms_identification_results_CSV \
-  --output_similarity_scores path_to_CSV_of_all_gcms_similarity_scores
 ```
 
-To implement an example of spectral library matching, you can navigate to the scripts directory and run the following commands which should produce CSV files of the identification results in your current working directory (i.e. the scripts directory) by default:
[…]
