### 2025-02-11 147c08f
Delete environment.yml
