+++ environment.yml
@@ -1,11 +1,0 @@
-name: mzsearch_env
-channels:
-  - conda-forge
-dependencies:
-  - python=3.12.4
-  - matplotlib=3.8.4
-  - numpy=1.26.4
-  - pandas=2.2.2
-  - scipy=1.13.1
-  - bioconda::pyteomics=4.7.2
-  - netCDF4=1.6.5
