### 2025-02-11 #7594 @hdlugas
Hello @aboffin  and @jromanowska, we just wanted to let you know that we have merged @aboffin's pull request into the main branch. We have also consolidated the LCMS and GCMS scripts into single scripts for (i) plotting spectra and (ii) running spectral library matching. Additionally, we removed the instructions for creating the conda environment and instead simply listed the dependencies, included the likely_gcms_ids.csv and likely_lcms_ids.csv files, and updated both the GitHub repository and manuscript to reflect these changes.

Regarding the help messages displayed with the -h flag, we were considering whether it might be cumbersome for users to reference the lengthy GitHub README file for parameter descriptions rather than simply examining them via the -h flag. We are unsure how to shorten these parameter descriptions without making the restrictions of each parameter unclear (e.g., the format of the query/reference libraries, the requirement that the entropy dimension must be positive, the nuances of the spectrum_preprocessing_order parameter, etc.), especially for users who do not thoroughly review the GitHub README file.

If you strongly recommend shortening the help messages, we will do so; however, if not, we would prefer to keep them as they are, as this may make it easier for users to decide which parameters to use without needing to switch between multiple windows (i.e. one with the terminal and another with the GitHub README containing more detailed descriptions of the parameters). Additionally, since they are only triggered when a user invokes the -h option, they do not necessarily add unnecessary clutter.
