### 2025-02-18 #7594 @jromanowska
👋🏼 @lionel42, thank you! As a reviewer, the first step is to create a checklist for your review by entering

```@editorialbot generate my checklist```

I encourage you to **over-communicate** and let everyone know that you're on the task every now and then, instead of waiting several weeks to collect all your comments, questions, or suggestions.

_Please feel free to ping me (@jromanowska) if you have any questions/concerns._
