### 2025-02-24 #7594 @lionel42
I have a question concerning the authors list.
I see that @hdlugas (Hunter Dlugas ) and author (Seongho Kim) contributed equally, but on the github insights (https://github.com/hdlugas/MZsearch/graphs/contributors) I see that the main coding contributions come from Hunter Dlugas who is the main author.
What is the justification to have these two authors as equal contributors ?
