### 2025-02-24 #7594 @lionel42
I have doubts about the point `Substantial scholarly effort`

If I look at the code, I see that most of it is about defining the command line arguments and functionality.

Most of the core functions (which are doing the calculations and are the interesting part) are included in `https://github.com/hdlugas/MZsearch/blob/main/src/processing.py` and in `https://github.com/hdlugas/MZsearch/blob/main/src/similarity_measures.py`

These two files are only around 300 lines and include mostly comments.
