### 2025-02-24 #7594 @lionel42
You mention the matchms package in python which is doing mostly the same as you (processing spectra + spectral matching + plotting).
In this case I have trouble to see how this package can do more than matchms, and wonder why not implementing the similarity measures directly in matchms ?
