### 2025-02-24 #7594 @lionel42
You need to add metadata about the data on your repository somewhere. Especially we need to know how/when it was collected and what was measured. This will help to decide point about: `Human and animal research`
