### 2025-02-24 #7594 @aboffin
> 👋 just checking up on this!
>
> [@aboffin](https://github.com/aboffin), [@lionel42](https://github.com/lionel42) - how is it going? Are you working on the reviews still?
>
> [@hdlugas](https://github.com/hdlugas) - are you working on improvements or waiting for comments from the reviewers?

@jromanowska I was a bit hesitant from the start about this submission, hence I made some preliminary review to make the code work and provided some comments to the author, I was planning to review the mass spec processing using my own data. However, as @lionel42 and I had pointed out, the scripts themselves appear to be mostly command-line options and documentation and if you take them away there is no substantial content. Either @hdlugas can contribute the code to matchms which does most of the things as MZsearch (echoing @lionel42's comment above) or make substantial changes to the code to make it a stand-alone and functional tool. @jromanowska what is the policy on AI-generated code, do the author(s) need to declare? Thank you!
