### 2025-02-25 #7594 @hdlugas
@jromanowska, @aboffin, @lionel42,

As you know, the key novelties of our package are twofold: (i) it allows users to customize the order of preprocessing steps, which significantly impacts compound identification accuracy, and (ii) it introduces two novel similarity measures - Tsallis and Renyi entropy similarity measures - that, to our knowledge, are not currently available in any other package, including matchms. While matchms also processes spectra and performs spectral matching, it requires users to write Python code to specify the order of the preprocessing steps. In contrast, our package allows users to specify preprocessing steps via the command line. This makes MZsearch more accessible to users who need fine-tuned spectral analysis who are more comfortable working in the command line than directly with Python.

To further enhance the accuracy of compound identification, we can add functionality for optimizing tuning parameters, including window sizes, weighting factors, thresholds for low-entropy transformation, and the q-values for Tsallis and Renyi entropy measures. This feature can allow users to refine parameter selection systematically and achieve improved identification performance, although would likely add at most a few hundred lines of code.
