### 2025-02-26 #7594 @lionel42
@hdlugas Thanks for your answers, so just to answer

> As you know, the key novelties of our package are twofold: (i) it allows users to customize the order of preprocessing steps, which significantly impacts compound identification accuracy

This is something most other matching packages can already do, I don't see how it is a novelty or important feature

> (ii) it introduces two novel similarity measures - Tsallis and Renyi entropy similarity measures - that, to our knowledge, are not currently available in any other package, including matchms.

 I still think that this would been better implemented in an existing framework

> In contrast, our package allows users to specify preprocessing steps via the command line. This makes MZsearch more accessible to users who need fine-tuned spectral analysis who are more comfortable working in the command line than directly with Python.

I don't know if this includes a lot of people in the community, but in general they prefer using well known and well maintained tools,  and I am not aware of anyone who prefers working with command line.

> To further enhance the accuracy of compound identification, we can add functionality for optimizing tuning parameters, including window sizes, weighting factors, thresholds for low-entropy transformation, and the q-values for Tsallis and Renyi entropy measures. This feature can allow users to refine parameter selection systematically and achieve improved identification performance, although would likely add at most a few hundred lines of code.

I think if your interest is to enhance accuracy of spectral matching methods, it would fit more in a traditional research paper, with results and discussion than on a software paper.

@jromanowska As I understand it is your responsibility to judge if the `Substantial scholarly effort` criterion is met. I have provided my thoughts about it and will proceed on the review if you think it is met.
