### 2025-03-03 #7594 @jromanowska
Thank you, @hdlugas and @lionel42, for your comments. I am not an expert in metabolomics, so I will need to rely on your judgement and arguments.

@hdlugas - it would help if you provide a table where you compare features of your software to the other Python libraries you mentioned in the paper.

@lionel42, @aboffin - I have gone through the code and I can see that a lot is about command-line interface and options. Do you think that providing CLI is not useful in your research area? Does using MZsearch require unnecessary switching between e.g., bash and python to perform full analysis of data?
