### 2025-03-06 #7594 @hdlugas
@jromanowska sure, a table comparing our tool to other relevant tools can be found here: [comparison_of_tools.docx](https://github.com/user-attachments/files/19101573/comparison_of_tools.docx). We can include this table in the main text if recommended as well.
