### 2025-03-06 #7594 @jromanowska
This is a great table! I'd recommend to include it in the paper - but without the "description" column and maybe without "repository" column, just create reference items instead.

@lionel42 , @aboffin - it seems from this table that MZsearch gives is a good addition to the tools available already. I'd think that having command-line interface instead of Python coding is good when using workflows. Many workflow software can handle switching between e.g., CLI and Python scripts.
