### 2025-03-06 #7594 @lionel42
@jromanowska @aboffin I want to point out here a confusion about `GCMS` vs `LCMS`

In you paper you say, line 40

```
In a GC-MS experiment, the resulting mass spectra typically have m/z values as integers. In40
contrast, LC-MS experiments produce mass spectra with high resolution m/z values, often41
with many decimal places.
```

This is wrong. The reason why you have mz values as integers or higher resolution depends only on the Mass spectrometer used.

So you should definitely change the vocabulary you use.

A more interesting comparison would be to distinguish the softwares that support MS1 and MS2 types.
