### 2025-03-06 #7594 @lionel42
@jromanowska I still don't think that the CLI is a useful addition in this area, as most researchers simply don't use that. They use GUIs or integrate the python code directly in their processing pipelines.

You ask about unecessary steps to perform full data analysis. Here is one with this programm:   you have to convert the data to a specific text format, which is a unnecessary step, and that most programs can use well known data formats in mass spectrometry (like mzml or mgf)
