### 2025-03-06 #7594 @aboffin
@hdlugas

1. In the Word document table, please clarify difference if any, between "spectral matching", "compound identification" and "annotation"?
2. Please respond to the other reviewer's comments regarding the author contributions. [Another author is listed but there was no github commits].
3. Can the program accept mzXML input and output (rather than providing a tab limited file)? It is trivial to generate such a file from mzXML but what are the implications downstream? How can users process the output further?
4. On a related note, please provide a workflow (or clarify in the text, the expected upstream and downstream steps i.e., where your program sits in a pipeline processing MS data).
5. Although not related to function, the code itself appears auto-generated or slightly modified boiler plate taken from examples. Code cleanup will substantially improve the quality and enable other users to contribute and improve the tool.
6. Please provide a comparison of test data analyzed with your program with similar tools, example utilizing similar models available in other tools.

@jromanowska
I have requested the author(s) to address a few issues. I do not have any reservation about CLI tools, I feel they definitely have their place in MS analyses. I am not sure objectively how I can evaluate if the contribution is significant or note at this moment without data from the authors. Also, as I have questions about the quality of the code and the generic responses that seem to address some of the reviewers' comments but seem to ignore certain other questions (for example, the contribution from the other author listed has not been answered).  Therefore, I request the authors to extensively revise the paper, code and address the reviewers' comments before any further consideration.
