### 2025-03-10 c7f1ff8
Update README.md
included is_reference flag in build_library.py
