### 2025-03-10 30961c0
Add files via upload
included is_reference flag for building a reference library with known compound IDs
