### 2025-03-10 5ff03b4
Update paper.md
include author contributions statement
