+++ paper/paper.md
@@ -206,4 +206,10 @@ Institutes of Health P20GM113226) and (National Institutes of Health
 P50AA024337). Computations were performed on Wayne State University's
 High-Performance Computing Grid.
 
+
+# Author Contributions
+
+S.K. conceived the study and contributed to the overall study design and interpretation. H.D. implemented the initial idea, expanded and improved upon it, and was responsible for coding the developed methods. Both S.K. and H.D. contributed equally to the manuscript preparation, including drafting and revising the paper. Given their significant and complementary contributions to both the conceptual and technical aspects of the study, we consider them equal contributors. X.Z. was involved in preliminary discussions and in editing the manuscript.
+
+
 # References
