+++ README.md
@@ -1,5 +1,5 @@
 # MZsearch
-Command-line Python tool to perform spectral library matching to identify chemical compounds with host of preprocessing transformations and similarity measures (Cosine and three entropy-based similarity measures). MZsearch is capable of performing spectral library matching with respect to either gas chromatography - mass spectrometry (GC-MS) or liquid chromatography - mass spectrometry (LC-MS) data.
+Command-line Python tool to perform spectral library matching to identify chemical compounds with host of preprocessing transformations and similarity measures (Cosine and three entropy-based similarity measures). MZsearch is capable of performing spectral library matching with respect to either gas chromatography - mass spectrometry (GC-MS) or liquid chromatography - mass spectrometry (LC-MS/MS) data.
 
 ## Table of Contents
 - [1. Install dependencies](#create-conda-env)
@@ -7,7 +7,7 @@ Command-line Python tool to perform spectral library matching to identify chemic
    - [2.1 Spectrum Preprocessing Transformations](#spec-preprocessing-transformations)
    - [2.2 Similarity Measures](#similarity-measures)
 - [3. Usage](#usage)
-   - [3.1 Obtain LC-MS or GC-MS library from MGF, mzML, or cdf file](#process-data)
+   - [3.1 Obtain LC-MS/MS or GC-MS library from MGF, mzML, or cdf file](#process-data)
    - [3.2 Run spectral library matching](#run-spec-lib-matching)
    - [3.3 Plot a query spectrum against a reference spectrum before and after spectrum preprocessing transformations](#plotting)
 - [4. Bugs/Questions?](#bugs-questions)
@@ -47,7 +47,7 @@ The following spectrum preprocessing transformations are offered:
     $x_{i}^{\star}=x_{i}^{\frac{1+H_{Shannon}(I)}{1+T}}$ if
     $H_{Shannon}(I)\textless T$.
 
--   Centroiding (only applicable to LC-MS data): Given a user-defined
+-   Centroiding (only applicable to LC-MS/MS data): Given a user-defined
     window-size parameter $w_{centroiding}$ and a spectrum $I$ with m/z
     values $(m_{1},m_{2},...,m_{n})$ and intensities
     $(x_{1},x_{2},...,x_{n})$, the transformed spectrum $I^{\star}$
@@ -65,7 +65,7 @@ The following spectrum preprocessing transformations are offered:
     $x_{j}\textless r\cdot\text{max}(\{x_{1},x_{2},...,x_{n}\})$ for
     $j\in\{1,2,...,n\}$.
 
--   Matching (only applicable to LC-MS data): Given a user-defined
+-   Matching (only applicable to LC-MS/MS data): Given a user-defined
     window-size parameter $w_{matching}$ and two spectra $I$, $J$ with
     m/z ratios $(a_{1},a_{2},...,a_{n}), (b_{1},b_{2},...,b_{m})$ and
     intensities $(x_{1},x_{2},...,x_{n}), (y_{1},y_{2},...,y_{m})$,
@@ -167,7 +167,7 @@ This repository has three main capabilities:
 2. running spectral library matching to identify compounds based off of their mass spectrometry data
 3. plotting a query spectrum vs a reference spectrum before and after preprocessing transformations.
 
-These tasks are implemented separately for the cases of (i) GC-MS and (ii) LC-MS data due to the different spectrum preprocessing transformations stemming from a different format in the mass to charge ratios in GC-MS vs LC-MS data. To see all parameters for any of the three main scripts (build_library.py, spec_lib_matching.py, plot_spectra.py), run:
[…]
