### 2025-03-11 09643f4
Update plot_spectra.py
