### 2025-03-11 02c145d
Update tests_plot_spectra.sh
