+++ tests/tests_plot_spectra.sh
@@ -8,7 +8,7 @@ python plot_spectra.py \
   --query_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_query_library.csv \
   --reference_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_reference_library.csv \
   --similarity_measure tsallis \
-  --chromatography_platform LCMS \
+  --chromatography_platform LCMSMS \
   --wf_mz 2 \
   --wf_intensity 0.5 \
   --normalization_method standard \
@@ -20,7 +20,7 @@ python plot_spectra.py \
   --query_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_query_library.csv \
   --reference_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_reference_library.csv \
   --similarity_measure tsallis \
-  --chromatography_platform LCMS \
+  --chromatography_platform LCMSMS \
   --wf_mz 2 \
   --wf_intensity 0.5 \
   --normalization_method standard \
@@ -32,7 +32,7 @@ python plot_spectra.py \
   --query_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_query_library.csv \
   --reference_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_reference_library.csv \
   --similarity_measure tsallis \
-  --chromatography_platform LCMS \
+  --chromatography_platform LCMSMS \
   --wf_mz 2 \
   --wf_intensity 0.5 \
   --normalization_method standard \
@@ -44,7 +44,7 @@ python plot_spectra.py \
   --query_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_query_library.csv \
   --reference_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_reference_library.csv \
   --similarity_measure tsallis \
-  --chromatography_platform LCMS \
+  --chromatography_platform LCMSMS \
   --normalization_method standard \
   --entropy_dimension 2 \
   --save_plots /home/hunter/mass_spec/entropies/JOSS/scripts/reviewer1/test3.pdf
@@ -54,7 +54,7 @@ python plot_spectra.py \
   --query_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_query_library.csv \
   --reference_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_reference_library.csv \
   --similarity_measure renyi \
-  --chromatography_platform LCMS \
+  --chromatography_platform LCMSMS \
   --normalization_method standard \
   --entropy_dimension 1.1 \
   --save_plots /home/hunter/mass_spec/entropies/JOSS/scripts/reviewer1/test4.pdf
@@ -65,7 +65,7 @@ python plot_spectra.py \
   --query_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_query_library.csv \
   --reference_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_reference_library.csv \
   --similarity_measure renyi \
-  --chromatography_platform LCMS \
+  --chromatography_platform LCMSMS \
   --normalization_method standard \
   --entropy_dimension 0.9 \
   --save_plots /home/hunter/mass_spec/entropies/JOSS/scripts/reviewer1/test5.pdf
@@ -75,7 +75,7 @@ python plot_spectra.py \
   --query_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_query_library.csv \
   --reference_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_reference_library.csv \
   --similarity_measure shannon \
[…]
