+++ tests/tests_spec_lib_matching.sh
@@ -7,7 +7,7 @@ echo $'\n\n\n\n\ntest #1'
 python spec_lib_matching.py \
   --query_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_query_library.csv \
   --reference_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_reference_library.csv \
-  --chromatography_platform LCMS \
+  --chromatography_platform LCMSMS \
   --similarity_measure tsallis \
   --wf_mz 2 \
   --wf_intensity 0.5 \
@@ -22,7 +22,7 @@ echo $'\n\n\n\n\ntest #2'
 python spec_lib_matching.py \
   --query_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_query_library.csv \
   --reference_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_reference_library.csv \
-  --chromatography_platform LCMS \
+  --chromatography_platform LCMSMS \
   --similarity_measure tsallis \
   --wf_mz 2 \
   --wf_intensity 0.5 \
@@ -35,7 +35,7 @@ echo $'\n\n\n\n\ntest #3'
 python spec_lib_matching.py \
   --query_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_query_library.csv \
   --reference_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_reference_library.csv \
-  --chromatography_platform LCMS \
+  --chromatography_platform LCMSMS \
   --similarity_measure tsallis \
   --normalization_method standard \
   --entropy_dimension 2 \
@@ -45,7 +45,7 @@ echo $'\n\n\n\n\ntest #4'
 python spec_lib_matching.py \
   --query_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_query_library.csv \
   --reference_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_reference_library.csv \
-  --chromatography_platform LCMS \
+  --chromatography_platform LCMSMS \
   --similarity_measure renyi \
   --normalization_method standard \
   --entropy_dimension 1.1 \
@@ -57,7 +57,7 @@ echo $'\n\n\n\n\ntest #5'
 python spec_lib_matching.py \
   --query_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_query_library.csv \
   --reference_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_reference_library.csv \
-  --chromatography_platform LCMS \
+  --chromatography_platform LCMSMS \
   --similarity_measure renyi \
   --normalization_method standard \
   --entropy_dimension 0.9 \
@@ -68,7 +68,7 @@ echo $'\n\n\n\n\ntest #6'
 python spec_lib_matching.py \
   --query_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_query_library.csv \
   --reference_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_reference_library.csv \
-  --chromatography_platform LCMS \
+  --chromatography_platform LCMSMS \
   --similarity_measure shannon \
   --normalization_method standard \
   --n_top_matches_to_save 2 \
@@ -78,7 +78,7 @@ echo $'\n\n\n\n\ntest #7'
 python spec_lib_matching.py \
   --query_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_query_library.csv \
   --reference_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_reference_library.csv \
-  --chromatography_platform LCMS \
+  --chromatography_platform LCMSMS \
   --similarity_measure cosine \
   --n_top_matches_to_save 2 \
[…]
