+++ tests/tests_spec_lib_matching.sh
@@ -1,12 +1,12 @@
 #!/bin/bash
 
-cd /home/hunter/mass_spec/entropies/JOSS/scripts/reviewer1
+cd ${PWD}/../src
 
 
 echo $'\n\n\n\n\ntest #1'
 python spec_lib_matching.py \
-  --query_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_query_library.csv \
-  --reference_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_reference_library.csv \
+  --query_data ../data/lcms_query_library.csv \
+  --reference_data ../data/lcms_reference_library.csv \
   --chromatography_platform LCMSMS \
   --similarity_measure tsallis \
   --wf_mz 2 \
@@ -15,13 +15,13 @@ python spec_lib_matching.py \
   --entropy_dimension 2 \
   --n_top_matches_to_save 4 \
   --print_id_results True \
-  --output_identification /home/hunter/mass_spec/entropies/JOSS/scripts/reviewer1/output_lcms_identification.csv \
-  --output_similarity_scores /home/hunter/mass_spec/entropies/JOSS/scripts/reviewer1/output_lcms_all_similarity_scores.csv
+  --output_identification ../scripts/reviewer1/output_lcms_identification.csv \
+  --output_similarity_scores ../scripts/reviewer1/output_lcms_all_similarity_scores.csv
 
 echo $'\n\n\n\n\ntest #2'
 python spec_lib_matching.py \
-  --query_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_query_library.csv \
-  --reference_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_reference_library.csv \
+  --query_data ../data/lcms_query_library.csv \
+  --reference_data ../data/lcms_reference_library.csv \
   --chromatography_platform LCMSMS \
   --similarity_measure tsallis \
   --wf_mz 2 \
@@ -33,8 +33,8 @@ python spec_lib_matching.py \
 
 echo $'\n\n\n\n\ntest #3'
 python spec_lib_matching.py \
-  --query_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_query_library.csv \
-  --reference_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_reference_library.csv \
+  --query_data ../data/lcms_query_library.csv \
+  --reference_data ../data/lcms_reference_library.csv \
   --chromatography_platform LCMSMS \
   --similarity_measure tsallis \
   --normalization_method standard \
@@ -43,8 +43,8 @@ python spec_lib_matching.py \
 
 echo $'\n\n\n\n\ntest #4'
 python spec_lib_matching.py \
-  --query_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_query_library.csv \
-  --reference_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_reference_library.csv \
+  --query_data ../data/lcms_query_library.csv \
+  --reference_data ../data/lcms_reference_library.csv \
   --chromatography_platform LCMSMS \
   --similarity_measure renyi \
   --normalization_method standard \
@@ -55,8 +55,8 @@ python spec_lib_matching.py \
 
 echo $'\n\n\n\n\ntest #5'
 python spec_lib_matching.py \
-  --query_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_query_library.csv \
-  --reference_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_reference_library.csv \
+  --query_data ../data/lcms_query_library.csv \
+  --reference_data ../data/lcms_reference_library.csv \
   --chromatography_platform LCMSMS \
[…]
