+++ tests/tests_plot_spectra.sh
@@ -1,106 +1,106 @@
 #!/bin/bash
 
-cd /home/hunter/mass_spec/entropies/JOSS/scripts/reviewer1
+cd ${PWD}/../src
 
 
 echo $'\n\n\n\n\ntest #1'
 python plot_spectra.py \
-  --query_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_query_library.csv \
-  --reference_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_reference_library.csv \
+  --query_data ../data/lcms_query_library.csv \
+  --reference_data ../data/lcms_reference_library.csv \
   --similarity_measure tsallis \
   --chromatography_platform LCMSMS \
   --wf_mz 2 \
   --wf_intensity 0.5 \
   --normalization_method standard \
   --entropy_dimension 2 \
-  --save_plots /home/hunter/mass_spec/entropies/JOSS/scripts/reviewer1/test1.pdf
+  --save_plots test1.pdf
 
 echo $'\n\n\n\n\ntest #2'
 python plot_spectra.py \
-  --query_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_query_library.csv \
-  --reference_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_reference_library.csv \
+  --query_data ../data/lcms_query_library.csv \
+  --reference_data ../data/lcms_reference_library.csv \
   --similarity_measure tsallis \
   --chromatography_platform LCMSMS \
   --wf_mz 2 \
   --wf_intensity 0.5 \
   --normalization_method standard \
   --entropy_dimension 2 \
-  --save_plots /home/hunter/mass_spec/entropies/JOSS/scripts/reviewer1/test2.pdf
+  --save_plots ../scripts/reviewer1/test2.pdf
 
 echo $'\n\n\n\n\ntest #2'
 python plot_spectra.py \
-  --query_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_query_library.csv \
-  --reference_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_reference_library.csv \
+  --query_data ../data/lcms_query_library.csv \
+  --reference_data ../data/lcms_reference_library.csv \
   --similarity_measure tsallis \
   --chromatography_platform LCMSMS \
   --wf_mz 2 \
   --wf_intensity 0.5 \
   --normalization_method standard \
   --entropy_dimension 0.5 \
-  --save_plots /home/hunter/mass_spec/entropies/JOSS/scripts/reviewer1/test2.pdf
+  --save_plots ../scripts/reviewer1/test2.pdf
 
 echo $'\n\n\n\n\ntest #3'
 python plot_spectra.py \
-  --query_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_query_library.csv \
-  --reference_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_reference_library.csv \
+  --query_data ../data/lcms_query_library.csv \
+  --reference_data ../data/lcms_reference_library.csv \
   --similarity_measure tsallis \
   --chromatography_platform LCMSMS \
   --normalization_method standard \
   --entropy_dimension 2 \
-  --save_plots /home/hunter/mass_spec/entropies/JOSS/scripts/reviewer1/test3.pdf
+  --save_plots ../scripts/reviewer1/test3.pdf
 
 echo $'\n\n\n\n\ntest #4'
 python plot_spectra.py \
-  --query_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_query_library.csv \
-  --reference_data /home/hunter/mass_spec/entropies/JOSS/data/lcms_reference_library.csv \
[…]
