+++ tests/tests_plot_spectra.sh
@@ -3,6 +3,18 @@
 cd ${PWD}/../src
 
 
+echo $'\n\n\n\n\ntest #0'
+python plot_spectra.py \
+  --query_data ../data/lcms_query_library.csv \
+  --reference_data ../data/lcms_reference_library.csv \
+  --similarity_measure tsallis \
+  --chromatography_platform LCMSMS \
+  --wf_mz 2 \
+  --wf_intensity 0.5 \
+  --normalization_method standard \
+  --entropy_dimension 2 \
+  --save_plots ../tests/test0.pdf
+
 echo $'\n\n\n\n\ntest #1'
 python plot_spectra.py \
   --query_data ../data/lcms_query_library.csv \
@@ -13,19 +25,7 @@ python plot_spectra.py \
   --wf_intensity 0.5 \
   --normalization_method standard \
   --entropy_dimension 2 \
-  --save_plots test1.pdf
-
-echo $'\n\n\n\n\ntest #2'
-python plot_spectra.py \
-  --query_data ../data/lcms_query_library.csv \
-  --reference_data ../data/lcms_reference_library.csv \
-  --similarity_measure tsallis \
-  --chromatography_platform LCMSMS \
-  --wf_mz 2 \
-  --wf_intensity 0.5 \
-  --normalization_method standard \
-  --entropy_dimension 2 \
-  --save_plots ../scripts/reviewer1/test2.pdf
+  --save_plots ../tests/test1.pdf
 
 echo $'\n\n\n\n\ntest #2'
 python plot_spectra.py \
@@ -37,7 +37,7 @@ python plot_spectra.py \
   --wf_intensity 0.5 \
   --normalization_method standard \
   --entropy_dimension 0.5 \
-  --save_plots ../scripts/reviewer1/test2.pdf
+  --save_plots ../tests/test2.pdf
 
 echo $'\n\n\n\n\ntest #3'
 python plot_spectra.py \
@@ -47,7 +47,7 @@ python plot_spectra.py \
   --chromatography_platform LCMSMS \
   --normalization_method standard \
   --entropy_dimension 2 \
-  --save_plots ../scripts/reviewer1/test3.pdf
+  --save_plots ../tests/test3.pdf
 
 echo $'\n\n\n\n\ntest #4'
 python plot_spectra.py \
@@ -57,7 +57,7 @@ python plot_spectra.py \
   --chromatography_platform LCMSMS \
   --normalization_method standard \
   --entropy_dimension 1.1 \
-  --save_plots ../scripts/reviewer1/test4.pdf
+  --save_plots ../tests/test4.pdf
 
 
 echo $'\n\n\n\n\ntest #5'
@@ -68,7 +68,7 @@ python plot_spectra.py \
   --chromatography_platform LCMSMS \
   --normalization_method standard \
   --entropy_dimension 0.9 \
-  --save_plots ../scripts/reviewer1/test5.pdf
+  --save_plots ../tests/test5.pdf
 
 echo $'\n\n\n\n\ntest #6'
 python plot_spectra.py \
@@ -77,7 +77,7 @@ python plot_spectra.py \
   --similarity_measure shannon \
   --chromatography_platform LCMSMS \
   --normalization_method standard \
-  --save_plots ../scripts/reviewer1/test6.pdf
+  --save_plots ../tests/test6.pdf
 
 echo $'\n\n\n\n\ntest #7'
 python plot_spectra.py \
@@ -85,7 +85,7 @@ python plot_spectra.py \
   --reference_data ../data/lcms_reference_library.csv \
   --similarity_measure cosine \
[…]
