+++ paper/paper.md
@@ -10,7 +10,7 @@ tags:
   - Cosine correlation
   - Shannon entropy
   - Renyi entropy
-  -  entropy
+  - entropy
 authors:
   - name: Hunter Dlugas
     orcid: 0000-0002-6819-0045
@@ -19,6 +19,7 @@ authors:
     orcid: 0000-0003-1102-6313
     affiliation: "3"
   - name: Seongho Kim
+    orcid: 0000-0003-1120-073X
     affiliation: "1, 2"
 affiliations:
  - name: Department of Oncology, School of Medicine, Wayne State University, Detroit, MI, USA
@@ -118,14 +119,15 @@ similarity measures: the commonly-used Cosine Similarity Measure
 recently-developed  Entropy Similarity Measure
 [@Dlugas2024_preprint], and the novel Rényi Entropy Similarity Measure. Table 1 compares MZsearch with other similar software.
 
-| Package  | Program language | User interface | MS/MS compatible | MS compatible | Similarity measures | Transformation(s) |
+| Package  | Program language | User interface | LC-MS/MS compatible | GC-MS compatible | Similarity measures | Preprocessing/Transformation(s) |
 |---------|:------:|:------:|:-----:|:-----:|:------:|:------:|
-| `MZsearch` | Python | command line | yes | yes | cosine, <br><br> Shannon, <br><br> Rényi, <br><br> Tsallis, | weight factor, <br><br> low-entropy, <br><br> filtering on m/z and intensity values, <br><br> noise removal |
-| `matchms` [@Huber2020] | Python | Python coding | yes | no | cosine-based, <br><br> Dice, <br><br> Jaccard, <br><br> custom | filtering on m/z and intensity values, <br><br> noise removal, <br><br> custom |
-| `OpenMS` [@Rost2014; @Rost2016] | C++/Python | C++/Python coding | yes | no | cosine | smoothing, <br><br> filtering, <br><br> other functionality not related to compound identification |
-| `spectrum_utils` [@Bittremieux2020] | Python | Python coding | yes | no | N/A | precursor & noise peak removal, <br><br> intensity filtering, <br><br> intensity scaling |
+| `MZsearch` | Python | command line; Python | yes | yes | cosine; <br><br> Shannon; <br><br> Rényi; <br><br> Tsallis | weight factor; <br><br> low-entropy; <br><br> filtering on m/z and intensity values; <br><br> noise removal |
+| `matchms` [@Huber2020] | Python | Python coding | yes | no | cosine-based; <br><br> Dice; <br><br> Jaccard; <br><br> custom | filtering on m/z and intensity values; <br><br> noise removal; <br><br> custom |
+| `OpenMS` [@Rost2014; @Rost2016] | C++/Python | C++/Python coding | yes | no | cosine | smoothing; <br><br> filtering; <br><br> other functionality not related to compound identification |
+| `spectrum_utils` [@Bittremieux2020] | Python | Python coding | yes | no | N/A | precursor & noise peak removal; <br><br> intensity filtering; <br><br> intensity scaling |
 | `metID` [@Shen2022] | R | R coding | yes | no | cosine | noise removal |
-| `ShinyMetID` [@Jeong2023] | R | R/Shiny GUI | no | yes | cosine, <br> weighted cosine, <br> Stein & Scott, <br><br> discrete Fourier transform, <br><br> discrete wavelet transform, <br><br> semi-partial correlation | weight factor
[…]
