+++ paper/paper.md
@@ -195,7 +195,7 @@ functionality for computing the following similarity measures:
 
 # Usage
 
-Compound identification occurs during preprocessing or as an independent procedure. In preprocessing, it follows peak detection and should be performed before cross-sample alignment if based on compound names; otherwise, it can be delayed until before pathway analysis. As an independent procedure, it can be conducted ad hoc or post hoc to confirm or identify specific peaks in single or multiple samples. MZsearch performs compound identification and has three main capabilities: (i) converting the raw data to the necessary format for spectral library matching, (ii) running spectral library matching to identify compounds based on their mass spectrometry data, and (iii) plotting a query spectrum vs. a reference spectrum before and after preprocessing transformations. These tasks are implemented separately for (i) GC-MS and (ii) LC-MS/MS data due to the different spectrum preprocessing transformations. To see all parameters for any of the three main scripts (build_library.py, spec_lib_matching.py, plot_spectra.py), see the documentation at <a href="[[url](https://github.com/hdlugas/MZsearch)] https://github.com/hdlugas/MZsearch)"></a>.
+Compound identification occurs during preprocessing or as an independent procedure. In preprocessing, it follows peak detection and should be performed before cross-sample alignment if based on compound names; otherwise, it can be delayed until before pathway analysis. As an independent procedure, it can be conducted ad hoc or post hoc to confirm or identify specific peaks in single or multiple samples. MZsearch performs compound identification and has three main capabilities: (i) converting the raw data to the necessary format for spectral library matching, (ii) running spectral library matching to identify compounds based on their mass spectrometry data, and (iii) plotting a query spectrum vs. a reference spectrum before and after preprocessing transformations. These tasks are implemented separately for (i) GC-MS and (ii) LC-MS/MS data due to the different spectrum preprocessing transformations. To see all parameters for any of the three main scripts (build_library.py, spec_lib_matching.py, plot_spectra.py), see the documentation at [https://github.com/hdlugas/MZsearch](https://github.com/hdlugas/MZsearch).
 
 
 # Acknowledgements
