+++ README.md
@@ -197,7 +197,7 @@ Parameter descriptions are as follows:
 
 Some example MGF files one can use to build a LC-MS/MS library can be found from the Global Natural Products Social Molecular Networking databases here: [https://external.gnps2.org/gnpslibrary](https://external.gnps2.org/gnpslibrary). Some example mzML file one can use to build a LC-MS/MS library can be found in this repository: [https://github.com/HUPO-PSI/mzML](https://github.com/HUPO-PSI/mzML). 
 
-Some GC-MS and LC-MS/MS reference libraries are available at the Zenodo database (Dlugas et al., 2024b). The reference libraries available in MZsearch are shortened versions (because GitHub offers limited storage) of these reference libraries found in the Zenodo database.
+Some GC-MS and LC-MS/MS reference libraries are available at the Zenodo database [https://zenodo.org/records/12786324](https://zenodo.org/records/12786324). The reference libraries available in MZsearch are shortened versions of these reference libraries due to GitHub's limited storage.
 
 <a name="run-spec-lib-matching"></a>
 ### 3.2 Run spectral library matching
