### 2025-03-20 #7594 @hdlugas
Hello @lionel42, @aboffin, @jromanowska, we collected all comments and suggestions, addressed them to the best of our ability, and revised MZsearch accordingly. The attached Word document contains our responses and comments. If you have any questions or suggestions regarding these changes or other aspects of MZsearch, please let us know.

[revision_comments.docx](https://github.com/user-attachments/files/19372765/revision_comments.docx)
