### 2025-03-20 #7594 @jromanowska
**lionel42**

**Q1**: I have a question concerning the authors list.

I see that \@hdlugas (Hunter Dlugas ) and author (Seongho Kim) contributed equally, but on the github insights (https://github.com/hdlugas/MZsearch/graphs/contributors) I see that the main coding contributions come from Hunter Dlugas who is the main author.

What is the justification to have these two authors as equal contributors ?

**Response**: In the previous version of the manuscript, we included a statement in the author list indicating equal contributions from H.D. and S.K. to reflect their significant and complementary contributions to both the conceptual and technical aspects of the study. However, following the reviewer's suggestion, we have removed this statement from the author list and instead included an \"Author Contributions\" section at the end of this revision, which reads as follows:

"*S.K. conceived the study and contributed to the overall study design and interpretation. H.D., X.Z., and S.K. collaborated on discussing the design, features, and structures of the developed software package. H.D. implemented the initial idea, further expanded and refined it, and was responsible for coding the developed software package. H.D., X.Z., and S.K. reviewed and edited the manuscript. Both H.D. and S.K. contributed equally to the manuscript preparation, including drafting and revising
the manuscript.*"

**Q2**: In you paper you say, line 40

In a GC-MS experiment, the resulting mass spectra typically have m/z values as integers. In contrast, LC-MS experiments produce mass spectra with high resolution m/z values, often with many decimal places.

This is wrong. The reason why you have mz values as integers or higher resolution depends only on the Mass spectrometer used.

So you should definitely change the vocabulary you use.

A more interesting comparison would be to distinguish the softwares that support MS1 and MS2 types.

**Response**: We understand that the resolution of m/z values depends on the type of mass spectrometer used. To clarify this point, we have revised the statement as follows:

\"*In many GC-MS experiments, the resulting mass spectra often have m/z values reported as integers due to the use of nominal mass resolution instruments. In contrast, LC-MS frequently employs high-resolution mass spectrometers that provide m/z values with multiple decimal places.*"

**Q3**: I still don\'t think that the CLI is a useful addition in this area, as most researchers simply don\'t use that. They use GUIs or integrate the python code directly in their processing pipelines.

You ask about unecessary steps to perform full data analysis.

Here is one with this programm: you have to convert the data to a specific text format, which is a unnecessary step, and that most programs can use well known data formats in mass spectrometry (like mzml or mgf)

**Response**: We appreciate the reviewer's feedback regarding the CLI approach and data format compatibility.

We acknowledge that many researchers prefer GUI-based tools or direct Python integration. However, a CLI-based tool offers distinct advantages, particularly in terms of automation and flexibility, making it well-suited for high-throughput processing pipelines. Additionally, CLI tools can be seamlessly integrated into Python workflows using standard functions such as 'subprocess.run()' to execute commands, and pandas or json to process structured results. This ensures accessibility for both coding-based users and those working with automated workflows.

For this reason, MZsearch was originally developed as a CLI-based tool. However, we recognize the reviewers' suggestions regarding direct Python integration. To improve usability, we have now included and documented instructions on using MZsearch within Python scripts. Specifically, we have added Section 3.2.2, \"Python Interface,\" to the README and
provided an example script, MZsearch/test/example\_code\_for\_python\_use.py, available on GitHub.
These additions ensure that users who prefer working directly within Python have a straightforward integration option.

At the same time, we recognize that requiring data conversion to a specific text format introduces an additional step. To minimize this inconvenience, we provide a conversion script within MZsearch to streamline the process. While mzML and MGF are widely used for full mass spectrometry datasets, they are not typically employed for storing individual mass spectra containing only m/z values and intensities. Instead, simpler formats such as CSV, TXT, or MSP are more commonly used
for this purpose. Our text-based format serves as a middle-ground solution, balancing the needs of researchers working on large-scale computational workflows with those focusing on validating the identification of a single compound. By allowing users to manually inspect and verify spectra for individual metabolites while also enabling automated processing for larger datasets, this approach ensures both accessibility and scalability.

Furthermore, MZsearch allows users to specify whether to build a reference library or a query library, offering greater flexibility in data preprocessing. This feature enables users to explore different orders of preprocessing transformations and similarity measures when evaluating compound identification results. Converting mzML, MGF, or CDF files to an intermediate format allows users to test different settings without requiring repeated raw data conversion. For those who need mzML
or MGF compatibility, existing tools such as pyOpenMS, pymzML, and msconvert provide conversion between formats, ensuring adaptability across different workflows. While we recognize the importance of direct support for standard MS formats, we believe that our current approach provides an accessible and practical solution for both small-scale
manual analysis and large-scale computational processing.

**Q4**: \@hdlugas You need to add metadata about the data on your repository somewhere. Especially we need to know how/when it was collected and what was measured. This will help to decide point about: Human and animal research

**Response**: All the data and their descriptions are available in the Zenodo database (<https://zenodo.org/records/12786324>). Following the reviewer's suggestion, we have included the relevant URL to full reference libraries in Section 3.1 of the repository's README file:

*"GC-MS and LC-MS/MS reference libraries are available at the Zenodo database (https://zenodo.org/records/12786324). The reference libraries available in MZsearch are shortened versions of these reference libraries due to GitHub\'s limited storage."*

Regarding the reviewer's question about species distinction, we note that metabolites' mass spectra are not systematically categorized by species, as spectral patterns are generally species-independent.

**Q5**: You mention the matchms package in python which is doing mostly the same as you (processing spectra + spectral matching + plotting).\ In this case I have trouble to see how this package can do more than matchms, and wonder why not implementing the similarity measures directly in matchms ?

**Response:** Thank you for the constructive question. At the request of the Editor, we have generated a comparison table, which is now included in the main text as 'Table 1. Comparison of MZsearch and Existing Tools.' This table compares and contrasts MZsearch with similar software
packages.

**\
**

**Aboffin**

**Q1**: In the Word document table, please clarify difference if any, between \"spectral matching\", \"compound identification\" and \"annotation\"?

**Response**: Thank you for the question. All these terms refer to compound identification. In the previous version of the table, we used the terminology as defined by each software package. However, in accordance with the Editor's suggestion, we did not include the description column in Table 1 in the revised manuscript.

**Q2**: Please respond to the other reviewer\'s comments regarding the author contributions. \[Another author is listed but there was no github commits\].

**Response**: We have addressed this in **Q1** of reviewer **lionel42**. For clarity, we have copied our response here:

In the previous version of the manuscript, we included a statement in the author list indicating equal contributions from H.D. and S.K. to reflect their significant and complementary contributions to both the conceptual and technical aspects of the study. However, following the reviewer's suggestion, we have removed this statement from the author list and instead included an \"Author Contributions\" section at the end of this revision, which reads as follows:

"*S.K. conceived the study and contributed to the overall study design and interpretation. H.D., X.Z., and S.K. collaborated on discussing the design, features, and structures of the developed software package. H.D. implemented the initial idea, further expanded and refined it, and was responsible for coding the developed software package. H.D., X.Z., and S.K. reviewed and edited the manuscript. Both H.D. and S.K. contributed equally to the manuscript preparation, including drafting and revising
the manuscript.*"

**Q3**: Can the program accept mzXML input and output (rather than providing a tab limited file)? It is trivial to generate such a file from mzXML but what are the implications downstream? How can users process the output further?

**Response**: MZsearch supports spectral library matching using CSV-formatted input, which can be generated from mzML, MGF, and CDF files using the provided conversion script. While generating a CSV file from mzXML is straightforward, we chose this intermediate format to ensure compatibility with multiple spectral formats and to offer greater flexibility for preprocessing steps such as filtering and transformation.

As we addressed in **Q3** of reviewer **lionel42**, while mzML and MGF are widely used for full mass spectrometry datasets, they are not typically employed for storing individual mass spectra that contain only m/z values and intensities. Instead, simpler formats such as CSV, TXT, or MSP are more commonly used for this purpose. Our text-based format
serves as a middle-ground solution, balancing the needs of researchers conducting large-scale computational workflows and those focusing on validating the identification of a single compound. By allowing users to manually inspect and verify spectra while also supporting automated processing for larger datasets, this approach ensures both accessibility and scalability.

Once identification results are obtained, users can perform a downstream analysis that summarizes the commonly identified compounds among various samples.

**Q4**: On a related note, please provide a workflow (or clarify in the text, the expected upstream and downstream steps i.e., where your program sits in a pipeline processing MS data).

**Response**: Compound identification is generally performed in two phases: as part of preprocessing in an MS data pipeline or as an independent procedure. In an MS data processing pipeline, compound identification typically follows peak detection, but its exact placement depends on subsequent analysis steps. If cross-sample alignment relies on compound names, identification should be performed before alignment. Otherwise, it can be delayed until before pathway analysis. As an independent procedure, compound identification can also be conducted ad hoc or post hoc to confirm or identify the names of specific peaks from a single or multiple samples. Following the reviewer's suggestion, we have added the following clarification to the 'Usage' section of the main text:

\"*Compound identification occurs during preprocessing or as an independent procedure. In preprocessing, it follows peak detection and should be performed before cross-sample alignment if based on compound names; otherwise, it can be delayed until before pathway analysis. As an independent procedure, it can be conducted ad hoc or post hoc to confirm
or identify specific peaks in single or multiple samples.*\"

**Q5**: Although not related to function, the code itself appears auto-generated or slightly modified boiler plate taken from examples. Code cleanup will substantially improve the quality and enable other users to contribute and improve the tool. Fix relative paths of test data, etc.

**Response**: The paths to the test data are structured to ensure that scripts can be run directly after cloning the repository, without requiring users to manually modify paths. This approach is preferable to using absolute paths specific to our local machine, as it allows users to run the toy example immediately, at least when paths are not explicitly specified. Regarding the reviewer's concern about the code quality, we acknowledge that improvements in code readability and
organization can facilitate community contributions. We are committed to code cleanup and refactoring where needed to enhance maintainability and usability. If the reviewer has specific suggestions for improvement, we would be happy to consider them in our next update. No AI tools were used in producing this code. If recommended, we would be glad to include this statement in the main text.

**Q6**: Please provide a comparison of test data analyzed with your program with similar tools, example utilizing similar models available in other tools.

**Response**: The primary contribution of our software is the implementation of new similarity measures and the ability to apply user-defined orders of spectral preprocessing transformations for compound identification. To address the reviewer's request for a feature comparison, we have included Table 1 in the revised manuscript, as suggested by the Editor, which outlines the differences between MZsearch and similar tools.

In terms of comparing test data analyzed with other tools, such a comparison would not be meaningful if the same similarity measures were used, as identical measures should yield the same identification performance across different tools. Since MZsearch introduces new similarity measures, a direct comparison between tools would not provide additional insights beyond what is already known about existing measures.

If the reviewer is referring to a comparison between our new similarity measures and previously established ones, this falls beyond the primary scope of our manuscript. Our focus is on presenting these new similarity measures as part of a software package, rather than benchmarking their performance across tools or preprocessing workflows. However, for those interested in a detailed evaluation of these measures, we refer to Li et al. (2021) and Dlugas et al. (2024), both of which provide comprehensive comparisons of Cosine, Shannon, and Tsallis similarity measures, but not Rényi similarity measure. These references are cited in our manuscript.

We hope this clarification sufficiently addresses the reviewer's concerns.
