### 2025-03-21 #7594 @lionel42
@jromanowska  Thanks for putting these comments here.

I am having very serious doubts about continuing this review, as the package and paper actually look better before addressing this first round of comments.

@hdlugas  Here are the main current issues:

* My comment Q2 was not understood. There are 3 main differences between GC-MS vs LC-MS,  low res (integer masses) vs highres (float masses) and MS1 vs MS2 (also called tandem mass spectromety, see https://en.wikipedia.org/wiki/Tandem_mass_spectrometry ). I have seen that many changes in the code have modified the `lcms` to `lcmsms` making us understand that MS2 is used with LC, and MS1 with GC. However as I tried to explain, the chromatographic (also called separation) part and the fragmentation part (high res vs low res and MS1 vs MS2) are **independant** . The current text (starting around line 30) in the paper is completely confusing and seems to have been written by someone that is very confused about the science behind. Also the data for LCMS provided does not correspond to the MS2 data type as the new `lcmsms` notation suggests
* In the updated table, I see now that matchms is not compatible with GC-MS, which is wrong
* The updated table contains the same confusion with GC-MS vs LC-MS instead of MS vs MS2 as I suggested in the previous comment
*  The python interface usage proposed is very wired (https://github.com/hdlugas/MZsearch?tab=readme-ov-file#322-python-interface). It does not follow any python good practice. (`*` character in the import, path to the code must be specified instead of installable package)
* I do not understand why @hdlugas does not want to use github to provide answers, answers which sound very "AI generated".
* The updated table is visually aweful which seems as if no one took care of having looked at it.

I hope what I say here is not violating the JOSS code of conduct, but I have a feeling to be facing an AI agent.
