### 2025-03-21 #7594 @aboffin
@jromanowska thank you for posting the reply to the reviewers'
comments from the author here.

- Removing the description from the table (as a response to question
  1), was never requested by the other reviewer, as far as I
  understand. In fact, a clear and accurate description helps the
  readers in understanding the application of the various tools listed
  in the table.
- In Question 5, I did not mention absolute path, I asked the authors
  to fix the relative paths so that the code can be run from the repo.
- Have all the authors reviewed the manuscript and consent to the
  content and changes made by the first author? Usually a statement:
  "All authors have reviewed the final version of the manuscript." or
  something similar is included in papers, not sure about JOSS's
  editorial policy on this.
- Questions about AI have been raised by both @lionel42 and me. In
  their response, they clarify that AI was not used to generate the
  code, however, this is the second or third time a partial,_"almost
  quite, but not entirely unlike"_ list of responses to reviewers'
  comments were provided that causes issues. Perhaps this avoidance of
  addressing all the comments is not intentional but it still is a
  cause for concern.
- I have other priorities for the next 2 weeks, I will be able to test
  the updated code and provide feedback by Apr 15.
  [@jroamnowska please let me know if this delay is acceptable.]
