+++ paper/paper.md
@@ -119,9 +119,7 @@ similarity measures: the commonly-used Cosine Similarity Measure
 recently-developed  Entropy Similarity Measure
 [@Dlugas2024_preprint], and the novel Rényi Entropy Similarity Measure. Table 1 compares MZsearch with other similar software.
 
-**Table 1.** Comparison of MZsearch and existing tools.
-
-!(table.PNG){width="100%," height="100%"}
+![Comparison of MZsearch and existing tools.(table.PNG){width="100%," height="100%"}
 
 # Functionality
 
