+++ README.md
@@ -1,5 +1,5 @@
 # MZsearch
-Command-line Python tool to perform spectral library matching to identify chemical compounds with host of spectrum preprocessing transformations and similarity measures (Cosine and three entropy-based similarity measures). MZsearch is capable of performing spectral library matching with respect to either gas chromatography - mass spectrometry (GC-MS) or liquid chromatography - tandem mass spectrometry (LC-MS/MS) data.
+Command-line Python tool to perform spectral library matching to identify chemical compounds with host of spectrum preprocessing transformations and similarity measures (Cosine and three entropy-based similarity measures). MZsearch is capable of performing spectral library matching with respect to either mass spectrometry (MS) or tandem mass spectrometry (MS/MS) data.
 
 ## Table of Contents
 - [1. Install dependencies](#create-conda-env)
@@ -49,7 +49,7 @@ The following spectrum preprocessing transformations are offered:
     $x_{i}^{\star}=x_{i}^{\frac{1+H_{Shannon}(I)}{1+T}}$ if
     $H_{Shannon}(I)\textless T$.
 
--   Centroiding (only applicable to LC-MS/MS data): Given a user-defined
+-   Centroiding (only applicable to HRMS data): Given a user-defined
     window-size parameter $w_{centroiding}$ and a spectrum $I$ with m/z
     values $(m_{1},m_{2},...,m_{n})$ and intensities
     $(x_{1},x_{2},...,x_{n})$, the transformed spectrum $I^{\star}$
@@ -67,7 +67,7 @@ The following spectrum preprocessing transformations are offered:
     $x_{j}\textless r\cdot\text{max}(\{x_{1},x_{2},...,x_{n}\})$ for
     $j\in\{1,2,...,n\}$.
 
--   Matching (only applicable to LC-MS/MS data): Given a user-defined
+-   Matching (only applicable to HRMS data): Given a user-defined
     window-size parameter $w_{matching}$ and two spectra $I$, $J$ with
     m/z ratios $(a_{1},a_{2},...,a_{n}), (b_{1},b_{2},...,b_{m})$ and
     intensities $(x_{1},x_{2},...,x_{n}), (y_{1},y_{2},...,y_{m})$,
@@ -169,7 +169,7 @@ This repository has three main capabilities:
 2. running spectral library matching to identify compounds based off of their mass spectrometry data
 3. plotting a query spectrum vs a reference spectrum before and after preprocessing transformations.
 
-These tasks are implemented separately for the cases of (i) GC-MS and (ii) LC-MS/MS data due to the different spectrum preprocessing transformations stemming from a different format in the mass to charge ratios in GC-MS vs LC-MS/MS data. To see all parameters for any of the three main scripts (build_library.py, spec_lib_matching.py, plot_spectra.py), run:
+These tasks are implemented separately for the cases of (i) LRMS and (ii) HRMS data due to the different spectrum preprocessing transformations stemming from a different format in the mass to charge ratios in LRMS vs HRMS data. To see all parameters for any of the three main scripts (build_library.py, spec_lib_matching.py, plot_spectra.py), run:
 ```
 python build_library.py -h
 python spec_lib_matching.py -h
@@ -177,7 +177,7 @@ python plot_spectra.py -h
 ```
 
 <a name="process-data"></a>
-### 3.1 Obtain LC-MS/MS or GC-MS library from MGF, mzML, or cdf file
+### 3.1 Obtain LC-MS/MS or -MS library from MGF, mzML, or cdf file
 To obtain a CSV file of LC-MS/MS spectra in the format necessary for spectral library matching from raw data in the form of an mgf, mzML, or cdf file, one can run:
 ```
 python build_library.py \
@@ -197,14 +197,14 @@ Parameter descriptions are as follows:
 
[…]
