+++ tests/tests_plot_spectra.sh
@@ -8,7 +8,7 @@ python plot_spectra.py \
   --query_data ../data/lcms_query_library.csv \
   --reference_data ../data/lcms_reference_library.csv \
   --similarity_measure tsallis \
-  --chromatography_platform LCMSMS \
+  --chromatography_platform HRMS \
   --wf_mz 2 \
   --wf_intensity 0.5 \
   --normalization_method standard \
@@ -20,7 +20,7 @@ python plot_spectra.py \
   --query_data ../data/lcms_query_library.csv \
   --reference_data ../data/lcms_reference_library.csv \
   --similarity_measure tsallis \
-  --chromatography_platform LCMSMS \
+  --chromatography_platform HRMS \
   --wf_mz 2 \
   --wf_intensity 0.5 \
   --normalization_method standard \
@@ -32,7 +32,7 @@ python plot_spectra.py \
   --query_data ../data/lcms_query_library.csv \
   --reference_data ../data/lcms_reference_library.csv \
   --similarity_measure tsallis \
-  --chromatography_platform LCMSMS \
+  --chromatography_platform HRMS \
   --wf_mz 2 \
   --wf_intensity 0.5 \
   --normalization_method standard \
@@ -44,7 +44,7 @@ python plot_spectra.py \
   --query_data ../data/lcms_query_library.csv \
   --reference_data ../data/lcms_reference_library.csv \
   --similarity_measure tsallis \
-  --chromatography_platform LCMSMS \
+  --chromatography_platform HRMS \
   --normalization_method standard \
   --entropy_dimension 2 \
   --save_plots ../tests/test3.pdf
@@ -54,7 +54,7 @@ python plot_spectra.py \
   --query_data ../data/lcms_query_library.csv \
   --reference_data ../data/lcms_reference_library.csv \
   --similarity_measure renyi \
-  --chromatography_platform LCMSMS \
+  --chromatography_platform HRMS \
   --normalization_method standard \
   --entropy_dimension 1.1 \
   --save_plots ../tests/test4.pdf
@@ -65,7 +65,7 @@ python plot_spectra.py \
   --query_data ../data/lcms_query_library.csv \
   --reference_data ../data/lcms_reference_library.csv \
   --similarity_measure renyi \
-  --chromatography_platform LCMSMS \
+  --chromatography_platform HRMS \
   --normalization_method standard \
   --entropy_dimension 0.9 \
   --save_plots ../tests/test5.pdf
@@ -75,7 +75,7 @@ python plot_spectra.py \
   --query_data ../data/lcms_query_library.csv \
   --reference_data ../data/lcms_reference_library.csv \
   --similarity_measure shannon \
-  --chromatography_platform LCMSMS \
+  --chromatography_platform HRMS \
   --normalization_method standard \
   --save_plots ../tests/test6.pdf
 
@@ -84,7 +84,7 @@ python plot_spectra.py \
   --query_data ../data/lcms_query_library.csv \
   --reference_data ../data/lcms_reference_library.csv \
   --similarity_measure cosine \
-  --chromatography_platform LCMSMS \
+  --chromatography_platform HRMS \
   --save_plots ../tests/test7.pdf
 
 echo $'\n\n\n\n\ntest #8'
@@ -93,7 +93,7 @@ python plot_spectra.py \
   --reference_data ../data/lcms_reference_library.csv \
   --spectrum_preprocessing_order LFWNCM \
   --similarity_measure cosine \
-  --chromatography_platform LCMSMS \
+  --chromatography_platform HRMS \
   --wf_mz 0.5 \
[…]
