+++ tests/tests_spec_lib_matching.sh
@@ -11,7 +11,7 @@ echo $'\n\n\n\n\ntest #1'
 python spec_lib_matching.py \
   --query_data ../data/lcms_query_library.csv \
   --reference_data ../data/lcms_reference_library.csv \
-  --chromatography_platform LCMSMS \
+  --chromatography_platform HRMS \
   --similarity_measure tsallis \
   --wf_mz 2 \
   --wf_intensity 0.5 \
@@ -26,7 +26,7 @@ echo $'\n\n\n\n\ntest #2'
 python spec_lib_matching.py \
   --query_data ../data/lcms_query_library.csv \
   --reference_data ../data/lcms_reference_library.csv \
-  --chromatography_platform LCMSMS \
+  --chromatography_platform HRMS \
   --similarity_measure tsallis \
   --wf_mz 2 \
   --wf_intensity 0.5 \
@@ -39,7 +39,7 @@ echo $'\n\n\n\n\ntest #3'
 python spec_lib_matching.py \
   --query_data ../data/lcms_query_library.csv \
   --reference_data ../data/lcms_reference_library.csv \
-  --chromatography_platform LCMSMS \
+  --chromatography_platform HRMS \
   --similarity_measure tsallis \
   --normalization_method standard \
   --entropy_dimension 2 \
@@ -49,7 +49,7 @@ echo $'\n\n\n\n\ntest #4'
 python spec_lib_matching.py \
   --query_data ../data/lcms_query_library.csv \
   --reference_data ../data/lcms_reference_library.csv \
-  --chromatography_platform LCMSMS \
+  --chromatography_platform HRMS \
   --similarity_measure renyi \
   --normalization_method standard \
   --entropy_dimension 1.1 \
@@ -61,7 +61,7 @@ echo $'\n\n\n\n\ntest #5'
 python spec_lib_matching.py \
   --query_data ../data/lcms_query_library.csv \
   --reference_data ../data/lcms_reference_library.csv \
-  --chromatography_platform LCMSMS \
+  --chromatography_platform HRMS \
   --similarity_measure renyi \
   --normalization_method standard \
   --entropy_dimension 0.9 \
@@ -72,7 +72,7 @@ echo $'\n\n\n\n\ntest #6'
 python spec_lib_matching.py \
   --query_data ../data/lcms_query_library.csv \
   --reference_data ../data/lcms_reference_library.csv \
-  --chromatography_platform LCMSMS \
+  --chromatography_platform HRMS \
   --similarity_measure shannon \
   --normalization_method standard \
   --n_top_matches_to_save 2 \
@@ -82,7 +82,7 @@ echo $'\n\n\n\n\ntest #7'
 python spec_lib_matching.py \
   --query_data ../data/lcms_query_library.csv \
   --reference_data ../data/lcms_reference_library.csv \
-  --chromatography_platform LCMSMS \
+  --chromatography_platform HRMS \
   --similarity_measure cosine \
   --n_top_matches_to_save 2 \
   --print_id_results True
@@ -91,7 +91,7 @@ echo $'\n\n\n\n\ntest #8'
 python spec_lib_matching.py \
   --query_data ../data/lcms_query_library.csv \
   --reference_data ../data/lcms_reference_library.csv \
-  --chromatography_platform LCMSMS \
+  --chromatography_platform HRMS \
   --spectrum_preprocessing_order LFWNCM \
   --similarity_measure cosine \
   --wf_mz 0.5 \
@@ -102,7 +102,7 @@ echo $'\n\n\n\n\ntest #9'
 python spec_lib_matching.py \
   --query_data ../data/lcms_query_library.csv \
   --reference_data ../data/lcms_reference_library.csv \
[…]
