### 2025-03-25 hdlugas/MZsearch#2 @lionel42
Handling of bad parameters
Hello,

I tried giving

```
 --chromatography_platform GCMS
```

to `spec_lib_matching.py`

This raises
```
NameError: name 'all_similarity_scores' is not defined
```

I also tried other things which did not raise or raise random errors.

```
--similarity_measure a
```

```
--spectrum_preprocessing_order FLMkasdf
```

Would need to implement handling of wrong values for these parameters

EDIT: it was a more general issue with bad values for many input parameters
