### 2025-03-26 #7594 @lionel42
Code quality was already discussed in this review.
Here are the points which should be addressed:

* Duplicate code in `spec_lib_matching.py` for the transformations which appeared to be very similar for HRMS and LRMS. Also in `test_similarity_measures.py`
* Poor exception handling when bad values are given see details (https://github.com/hdlugas/MZsearch/issues/2)
* Poor usage of the argparse module. https://docs.python.org/3/library/argparse.html The argparse module lets developper add many custom info on variables such as types or default values. In this package (which claims that one of the main improvement to other packages is the CLI) only a very basic usage is done (the help).
* Usage of inline comments instead of docstings, as commonly done in python
