### 2025-03-26 #7594 @lionel42
@jromanowska I have been able to test the code. I updated the checklist above, as you can see there are many points still missing. I am happy to explain more if necessary.
