### 2025-03-27 71277e1
included more thorough exception handling and converted multi-line comments with pounds signs to docstrings
