+++ src/similarity_measures.py
@@ -1,6 +1,6 @@
 
 ##### Similarity Score Functions #####
-# Note that the input for all similarity measures are two 1-d arrays of the same length. 
+# Note that the input for all similarity measures are two 1-d np arrays of the same length. 
 # These 1-d arrays must be normalized to sum to 1 for the Shannon, Renyi, and Tsallis Entropy Similarity Measures.
 
 import scipy.stats
@@ -8,27 +8,33 @@ import numpy as np
 
 
 def S_cos(ints_a, ints_b):
-    #Cosine Similarity Measure
+    # Cosine Similarity Measure
     if np.sum(ints_a) == 0 or np.sum(ints_b) == 0:
         return(0)
     else:
         return np.dot(ints_a,ints_b) / (np.sqrt(sum(np.power(ints_a,2))) * np.sqrt(sum(np.power(ints_b,2))))
 
+
 def ent_renyi(ints, q):
-    #Computes the Renyi entropy of a probability distribution for a given positive entropy dimension q
+    # Computes the Renyi entropy of a probability distribution for a given positive entropy dimension q
     return np.log(sum(np.power(ints,q))) / (1-q)
 
+
 def ent_tsallis(ints, q):
-    #Computes the Tsallis entropy of a probability distribution for a given positive entropy dimension q
+    # Computes the Tsallis entropy of a probability distribution for a given positive entropy dimension q
     return (sum(np.power(ints,q))-1) / (1-q)
 
 
 def S_shannon(ints_a, ints_b):
-    #Shannon Entropy Similarity Measure
-    #This similarity function was presented by: 
-    #Li, Y.; Kind, T.; Folz, J.; Vaniya, A.; Mehta, S. S.; Fiehn, O.
-    #Spectral entropy outperforms MS/MS dot product similarity for small-molecule compound identification. 
-    #Note that since scipy.stats.entropy normalizes the input vector to sum to 1, vec1 and vec1 need not be normalized when computing ent_ab
+    '''
+    Shannon Entropy Similarity Measure
+
+    This similarity function was presented by: 
+    Li, Y.; Kind, T.; Folz, J.; Vaniya, A.; Mehta, S. S.; Fiehn, O.
+    Spectral entropy outperforms MS/MS dot product similarity for small-molecule compound identification. 
+    * Note that since scipy.stats.entropy normalizes the input vector to sum to 1, vec1 and vec1 need not be normalized when computing ent_ab
+    '''
+
     ent_a = scipy.stats.entropy(ints_a)
     ent_b = scipy.stats.entropy(ints_b)
     ent_ab = scipy.stats.entropy(ints_a + ints_b)
@@ -36,10 +42,12 @@ def S_shannon(ints_a, ints_b):
 
 
 def S_renyi(ints_a, ints_b, q):
-    #Renyi Entropy Similarity Measure
-    #This is a novel similarity measure which generalizes the Shannon Entropy Similarity Measure
-    #The Renyi Similarity Measure approaches the Shannon Entropy Similiarity Measure as q approaches 1
-    #ints_a and ints_b must be normalized to sum to 1
+    '''
+    Renyi Entropy Similarity Measure
+    * This is a novel similarity measure which generalizes the Shannon Entropy Similarity Measure
+    * The Renyi Similarity Measure approaches the Shannon Entropy Similiarity Measure as q approaches 1
+    * ints_a and ints_b must be normalized to sum to 1
+    '''
     if q == 1:
[…]
