+++ src/processing.py
@@ -1,40 +1,44 @@
 
-#This script contains the functions used to transform spectra prior to computing similarity scores
+# This script contains the functions used to transform spectra prior to computing similarity scores
 
 import scipy.stats
 import numpy as np
 
 def wf_transform(spec_mzs, spec_ints, wf_mz, wf_int):
-    #This function performs a weight factor transformation on a spectrum
+    '''
+    This function performs a weight factor transformation on a spectrum
     
-    #input:
-    #wf_int: float
-    #wf_mz: float
-    #spec_mzs: 1d np array representing mass/charge values 
-    #spec_ints: 1d np array representing intensity values 
+    input:
+    wf_int: float
+    wf_mz: float
+    spec_mzs: 1d np array representing mass/charge values 
+    spec_ints: 1d np array representing intensity values 
 
-    #spec_mzs and spec_ints must be of the same length N
+    spec_mzs and spec_ints must be of the same length N
 
-    #output:
-    #spec_ints: 1d np array of weight-factor-transformed spectrum intensities
+    output:
+    spec_ints: 1d np array of weight-factor-transformed spectrum intensities
+    '''
 
     spec_ints = np.power(spec_mzs, wf_mz) * np.power(spec_ints, wf_int)
     return(spec_ints)
 
 
 def LE_transform(intensity, thresh, normalization_method):
-    #This transformation was presented by: 
-    #Li, Y.; Kind, T.; Folz, J.; Vaniya, A.; Mehta, S. S.; Fiehn, O.
-    #Spectral entropy outperforms MS/MS dot product similarity for small-molecule compound identification. 
-    #Nature Methods 2021, 18, 1524–1531
+    '''
+    This transformation was presented by: 
+    Li, Y.; Kind, T.; Folz, J.; Vaniya, A.; Mehta, S. S.; Fiehn, O.
+    Spectral entropy outperforms MS/MS dot product similarity for small-molecule compound identification. 
+    Nature Methods 2021, 18, 1524–1531
     
-    #input:
-    #intensity: 1d np array
-    #thresh: nonnegative float
-    #normalization_method: either 'standard' or 'softmax'
+    input:
+    intensity: 1d np array
+    thresh: nonnegative float
+    normalization_method: either 'standard' or 'softmax'
     
-    #output:
-    #1d np array of transformed intensities
+    output:
+    1d np array of transformed intensities
+    '''
 
     intensity_tmp = normalize(intensity, method=normalization_method)
     if np.sum(intensity_tmp) > 0:
@@ -48,7 +52,17 @@ def LE_transform(intensity, thresh, normalization_method):
 
 
 def normalize(intensities,method='standard'):
-    #Normalizes a given vector to sum to 1 so that it represents a probability distribution
+    '''
+    Normalizes a given vector to sum to 1 so that it represents a probability distribution
+
+    input:
+    intensities: 1d np array
+    method: normalization method; either 'standard' or 'softmax'
+
+    output:
+    1d np array of normalized intensities
+    '''
+
     if np.sum(intensities) > 0:
         if method == 'softmax':
             intensities2 = np.exp(intensities)
@@ -62,7 +76,20 @@ def normalize(intensities,method='standard'):
 
 
 def filter_spec_lcms(spec, mz_min = 0, mz_max = 999999999999, int_min = 0, int_max = 999999999999, is_matched = False):
-    #keep points in a given spectrum in a given range of mz values and intensity values
+    '''
+    keep points in a given spectrum in a given range of mz values and intensity values
+
+    input:
[…]

[…]
