### 2025-03-27 hdlugas/MZsearch#2 @hdlugas
More comprehensive exception handling has now been included to handle cases such as these.
