### 2025-04-01 4a6bac7
incorporated argparse\'s functionality of specifying type and default values for CLI-input
