+++ tests/tests_plot_spectra.sh
@@ -358,7 +358,7 @@ echo $'\n\n\n\n\ntest #32'
 python plot_spectra.py \
   --query_data ../data/gcms_query_library.csv \
   --reference_data ../data/gcms_reference_library.csv \
-  --chromatography_platform LRMS \
+  --chromatography_platform NRMS \
   --similarity_measure cosine \
   --normalization_method standard \
   --save_plots ../tests/test32.pdf
@@ -367,7 +367,7 @@ echo $'\n\n\n\n\ntest #33'
 python plot_spectra.py \
   --query_data ../data/gcms_query_library.csv \
   --reference_data ../data/gcms_reference_library.csv \
-  --chromatography_platform LRMS \
+  --chromatography_platform NRMS \
   --similarity_measure tsallis \
   --wf_mz 2 \
   --wf_intensity 0.5 \
@@ -379,7 +379,7 @@ echo $'\n\n\n\n\ntest #34'
 python plot_spectra.py \
   --query_data ../data/gcms_query_library.csv \
   --reference_data ../data/gcms_reference_library.csv \
-  --chromatography_platform LRMS \
+  --chromatography_platform NRMS \
   --similarity_measure tsallis \
   --wf_mz 2 \
   --wf_intensity 0.5 \
@@ -391,7 +391,7 @@ echo $'\n\n\n\n\ntest #35'
 python plot_spectra.py \
   --query_data ../data/gcms_query_library.csv \
   --reference_data ../data/gcms_reference_library.csv \
-  --chromatography_platform LRMS \
+  --chromatography_platform NRMS \
   --similarity_measure tsallis \
   --normalization_method standard \
   --entropy_dimension 2 \
@@ -401,7 +401,7 @@ echo $'\n\n\n\n\ntest #36'
 python plot_spectra.py \
   --query_data ../data/gcms_query_library.csv \
   --reference_data ../data/gcms_reference_library.csv \
-  --chromatography_platform LRMS \
+  --chromatography_platform NRMS \
   --similarity_measure renyi \
   --normalization_method standard \
   --entropy_dimension 1.1 \
@@ -411,7 +411,7 @@ echo $'\n\n\n\n\ntest #37'
 python plot_spectra.py \
   --query_data ../data/gcms_query_library.csv \
   --reference_data ../data/gcms_reference_library.csv \
-  --chromatography_platform LRMS \
+  --chromatography_platform NRMS \
   --similarity_measure renyi \
   --normalization_method standard \
   --entropy_dimension 0.9 \
@@ -421,7 +421,7 @@ echo $'\n\n\n\n\ntest #38'
 python plot_spectra.py \
   --query_data ../data/gcms_query_library.csv \
   --reference_data ../data/gcms_reference_library.csv \
-  --chromatography_platform LRMS \
+  --chromatography_platform NRMS \
   --similarity_measure shannon \
   --normalization_method standard \
   --save_plots ../tests/test38.pdf
@@ -430,7 +430,7 @@ echo $'\n\n\n\n\ntest #39'
 python plot_spectra.py \
   --query_data ../data/gcms_query_library.csv \
   --reference_data ../data/gcms_reference_library.csv \
-  --chromatography_platform LRMS \
+  --chromatography_platform NRMS \
   --similarity_measure cosine \
   --save_plots ../tests/test39.pdf
 
@@ -438,7 +438,7 @@ echo $'\n\n\n\n\ntest #40'
 python plot_spectra.py \
   --query_data ../data/gcms_query_library.csv \
   --reference_data ../data/gcms_reference_library.csv \
-  --chromatography_platform LRMS \
+  --chromatography_platform NRMS \
[…]
