+++ tests/tests_spec_lib_matching.sh
@@ -3,6 +3,7 @@
 
 cd ${PWD}/../src
 
+
 echo $'\n\n\n\n\ntest #0'
 python spec_lib_matching.py \
   --query_data ../data/lcms_query_library.csv \
@@ -30,7 +31,7 @@ echo $'\n\n\n\n\ntest #3'
 python spec_lib_matching.py \
   --query_data ../data/lcms_query_library.csv \
   --reference_data ../data/lcms_reference_library.csv \
-  --chromatography_platform LRMS \
+  --chromatography_platform NRMS \
   --spectrum_preprocessing_order MF
 
 
@@ -427,7 +428,7 @@ echo $'\n\n\n\n\ntest #39'
 python spec_lib_matching.py \
   --query_data ../data/gcms_query_library.csv \
   --reference_data ../data/gcms_reference_library.csv \
-  --chromatography_platform LRMS \
+  --chromatography_platform NRMS \
   --similarity_measure cosine \
   --normalization_method standard \
   --print_id_results True \
@@ -437,7 +438,7 @@ echo $'\n\n\n\n\ntest #40'
 python spec_lib_matching.py \
   --query_data ../data/gcms_query_library.csv \
   --reference_data ../data/gcms_reference_library.csv \
-  --chromatography_platform LRMS \
+  --chromatography_platform NRMS \
   --similarity_measure tsallis \
   --wf_mz 2 \
   --wf_intensity 0.5 \
@@ -453,7 +454,7 @@ echo $'\n\n\n\n\ntest #41'
 python spec_lib_matching.py \
   --query_data ../data/gcms_query_library.csv \
   --reference_data ../data/gcms_reference_library.csv \
-  --chromatography_platform LRMS \
+  --chromatography_platform NRMS \
   --similarity_measure tsallis \
   --wf_mz 2 \
   --wf_intensity 0.5 \
@@ -467,7 +468,7 @@ echo $'\n\n\n\n\ntest #42'
 python spec_lib_matching.py \
   --query_data ../data/gcms_query_library.csv \
   --reference_data ../data/gcms_reference_library.csv \
-  --chromatography_platform LRMS \
+  --chromatography_platform NRMS \
   --similarity_measure tsallis \
   --normalization_method standard \
   --entropy_dimension 2 \
@@ -478,7 +479,7 @@ echo $'\n\n\n\n\ntest #43'
 python spec_lib_matching.py \
   --query_data ../data/gcms_query_library.csv \
   --reference_data ../data/gcms_reference_library.csv \
-  --chromatography_platform LRMS \
+  --chromatography_platform NRMS \
   --similarity_measure renyi \
   --normalization_method standard \
   --entropy_dimension 1.1 \
@@ -490,7 +491,7 @@ echo $'\n\n\n\n\ntest #44'
 python spec_lib_matching.py \
   --query_data ../data/gcms_query_library.csv \
   --reference_data ../data/gcms_reference_library.csv \
-  --chromatography_platform LRMS \
+  --chromatography_platform NRMS \
   --similarity_measure renyi \
   --normalization_method standard \
   --entropy_dimension 0.9 \
@@ -502,7 +503,7 @@ echo $'\n\n\n\n\ntest #45'
 python spec_lib_matching.py \
   --query_data ../data/gcms_query_library.csv \
   --reference_data ../data/gcms_reference_library.csv \
-  --chromatography_platform LRMS \
+  --chromatography_platform NRMS \
   --similarity_measure shannon \
   --normalization_method standard \
   --n_top_matches_to_save 2 \
@@ -513,7 +514,7 @@ echo $'\n\n\n\n\ntest #46'
 python spec_lib_matching.py \
   --query_data ../data/gcms_query_library.csv \
[…]
