+++ README.md
@@ -1,5 +1,5 @@
 # MZsearch
-Command-line Python tool to perform spectral library matching to identify chemical compounds with host of spectrum preprocessing transformations and similarity measures (Cosine and three entropy-based similarity measures). MZsearch is capable of performing spectral library matching with respect to either nominal resolution mass spectrometry (NRMS) data (e.g., GC-MS) or high resolution mass spectrometry (HRMS) data (e.g., LC-MS/MS).
+A Python-based tool for spectral library matching, MZsearch is available in two versions: a command-line interface and Python modules for integration into customer code. It performs spectral library matching to identify chemical compounds, offering a range of spectrum preprocessing transformations and similarity measures, including Cosine and three entropy-based similarity measures. MZsearch supports both high-resolution mass spectrometry (HRMS) data (e.g., LC-MS/MS) and nominal-resolution mass spectrometry (NRMS) data (e.g., GC-MS).
 
 ## Table of Contents
 - [1. Install dependencies](#create-conda-env)
@@ -16,7 +16,7 @@ Command-line Python tool to perform spectral library matching to identify chemic
 
 <a name="create-conda-env"></a>
 ## 1. Install dependencies
-MZsearch requires the Python dependencies Matplotlib, NumPy, Pandas, and SciPy, Pyteomics, and netCDF4. Specifically, this software was validated with python=3.12.4, matplotlib=3.8.4, numpy=1.26.4, pandas=2.2.2, scipy=1.13.1, pyteomics=4.7.2, and netCDF4=1.6.5, although it may work with other versions of these tools. A user may consider creating a conda environment (see [https://docs.conda.io/projects/conda/en/latest/user-guide/getting-started.html](https://docs.conda.io/projects/conda/en/latest/user-guide/getting-started.html) for guidance on getting started with conda if you are unfamiliar).
+MZsearch requires the Python dependencies Matplotlib, NumPy, Pandas, SciPy, Pyteomics, and netCDF4. Specifically, this software was validated with python=3.12.4, matplotlib=3.8.4, numpy=1.26.4, pandas=2.2.2, scipy=1.13.1, pyteomics=4.7.2, and netCDF4=1.6.5, although it may work with other versions of these tools. A user may consider creating a conda environment (see [https://docs.conda.io/projects/conda/en/latest/user-guide/getting-started.html](https://docs.conda.io/projects/conda/en/latest/user-guide/getting-started.html) for guidance on getting started with conda if you are unfamiliar).
 
 <a name="functionality"></a>
 ## 2. Functionality
@@ -58,7 +58,7 @@ The following spectrum preprocessing transformations are offered:
     $(\frac{m_{i}\cdot x_{i}+m_{i+1}\cdot x_{i+1}}{x_{i}+x_{i+1}},x_{i}+x_{i+1})$
     if $|m_{i}-m_{i+1}|\textless w_{centroiding}$ for
     $i\in\{1,2,...,n-1\}$. This centroiding procedure generalizes to
-    more than two peaks whose m/z values are within distance
+    more than two peaks whose m/z values are within a distance
     $w_{centroiding}$ of each other.
 
 -   Noise Removal: Given a user-defined noise removal parameter $r$ and
@@ -165,11 +165,11 @@ q\neq 1, \ q\textgreater 0
 <a name="usage"></a>
 ## 3. Usage
 This repository has three main capabilities:
-1. converting the raw data to the necessary format for spectral library matching
-2. running spectral library matching to identify compounds based off of their mass spectrometry data
-3. plotting a query spectrum vs a reference spectrum before and after preprocessing transformations.
+1. Converting raw data to the necessary format for spectral library matching
[…]
