+++ README.md
@@ -314,6 +314,7 @@ from similarity_measures import *
 
 In particular, the available spectrum preprocessing transformations and similarity measures are:
 ```
+# Weight factor transformation
 wf_transform(spec_mzs, spec_ints, wf_mz, wf_int)
 """
 Perform weight factor transformation on a spectrum
@@ -326,122 +327,148 @@ Returns:
    np.ndarray: 1d numpy array of weight-factor-transformed spectrum intensities
 """
 
-LE_transform(intensity, thresh, normalization_method)  # transforms spectrum's intensities if the Shannon entropy of the intensities is below some threshold
-##### input: #####
-# intensity: 1d numpy array
-# thresh: nonnegative float
-# normalization_method: either 'standard' or 'softmax'
+# Low-entropy transformation
+LE_transform(intensity, thresh, normalization_method)
+"""
+Transforms spectrum's intensities if the Shannon entropy of the intensities is below some threshold
+Args:
+   intensity: 1d numpy array
+   thresh: nonnegative float
+   normalization_method: either 'standard' or 'softmax'
+Returns:
+   np.ndarray: 1d numpy array of transformed intensities
+"""
 
-##### output: #####
-# 1d numpy array of transformed intensities
+# Normalize intensities to sum to 1
+normalize(intensities, method)
+"""
+Normalize a spectrum so intensities sum to 1 so that the intensities represent a probability distribution
+Args:
+   intensities: 1d numpy array
+   method: either 'standard' or 'softmax'
+Returns:
+   np.ndarray: 1d numpy array of normalized intensities
+"""
 
+# Filter HR-MS such as LC-MS/MS spectrum
+filter_spec_lcms(spec, mz_min, mz_max, int_min, int_max, is_matched)
+"""
+Filter an MS/MS spectrum based on m/z and intensity values
+Args:
+   spec: N x 2 numpy array with first column being m/z and second column being intensity
+   mz_min: minimum m/z value
+   mz_max: maximum m/z value
+   int_min: minimum intensity value
+   int_max: maximum intensity value
+   is_matched: flag to indicate whether the given spectrum has already been matched to another spectrum
+Returns:
+   np.ndarray: N x 2 numpy array with intensity of 0 put anywhere outside of the m/z and/or intensity bounds
+"""
 
-normalize(intensities, method)  # normalize a spectrum so intensities sum to 1 so that the intensities represent a probability distribution
-##### input: #####
-# intensities: 1d numpy array
-# method: either 'standard' or 'softmax'
-
-##### output: #####
-# 1d numpy array of normalized intensities
-
-
-filter_spec_lcms(spec, mz_min, mz_max, int_min, int_max, is_matched)  # filter an MS/MS spectrum based on m/z and intensity values
-##### input: #####
-# spec: N x 2 numpy array with first column being m/z and second column being intensity
-# mz_min: minimum m/z value
-# mz_max: maximum m/z value
-# int_min: minimum intensity value
-# int_max: maximum intensity value
-# is_matched: flag to indicate whether the given spectrum has already been matched to another spectrum
-
-##### output: #####
-# N x 2 numpy array with intensity of 0 put anywhere outside of the m/z and/or intensity bounds
-
-
+# Filter NR-MS such as GC-MS spectrum
 filter_spec_gcms(spec, mz_min, mz_max, int_min, int_max)
-# spec: N x 2 numpy array with first column being m/z and second column being intensity
-# mz_min: minimum m/z value
-# mz_max: maximum m/z value
-# int_min: minimum intensity value
-# int_max: maximum intensity value
+"""
+Filter an MS spectrum based on m/z and intensity values
+Args:
[…]
