+++ paper/paper.md
@@ -37,74 +37,15 @@ editor_options:
 
 # Summary
 
-A primary goal in metabolomics, the qualitative and quantitative study
-of metabolites, is to identify chemical compounds in a given sample.
-This often involves identifying and quantifying metabolic biomarkers for
-use in diagnosing, treating, and stratifying the risk of various
-diseases. A central tool for compound identification in metabolomics is
-mass spectrometry, which produces the mass spectrum for a given chemical
-compound. Each peak in the mass spectrum represents an ion fragment,
-with one axis representing mass to charge ratio (m/z) and the other axis
-representing relative intensity. A popular method used to identify an
-unknown chemical compound based on its mass spectrum is spectral library
-matching. This process involves computing a measure of similarity
-between the mass spectrum of the unknown chemical compound and each mass
-spectrum in a reference library. The unknown chemical compound is then
-identified as the one from the reference library whose mass spectrum is
-most similar to that of the unknown compound. We present MZsearch, a
-command-line tool written in Python that implements spectral library
-matching. The MZsearch tool offers a range of spectrum preprocessing
-transformations and similarity measures, and it is capable of analyzing
-data generated from gas chromatography-mass spectrometry (GC-MS) and
-liquid chromatography-mass spectrometry (LC-MS/MS).
+A primary goal in metabolomics, the qualitative and quantitative study of metabolites, is to identify chemical compounds in a given sample. This often involves identifying and quantifying metabolic biomarkers for use in diagnosing, treating, and stratifying the risk of various diseases. A central tool for compound identification in metabolomics is mass spectrometry, which produces the mass spectrum for a given chemical compound. Each peak in the mass spectrum represents an ion fragment, with one axis representing mass-to-charge ratio (m/z) and the other axis representing relative intensity. A popular method used to identify a compound based on its mass spectrum is spectral library matching. This process involves computing a measure of similarity between the mass spectrum of the compound and each mass spectrum in a reference library. The compound is then identified as the one from the reference library whose mass spectrum is most similar to that of the compound. We present MZsearch, a Python-based tool for spectral library matching, available in two versions: a command-line interface and Python modules for integration into custom code. The MZsearch tool offers a range of spectrum preprocessing transformations and similarity measures and can identify compounds from nominal-resolution mass spectra (NR-MS) and high-resolution mass spectra (HR-MS).
 
 # Statement of need
 
-There exist several open-source software tools for working with mass
-spectrometry data. The Python package matchms allows one to perform
-spectral library matching for both GC-MS and LC-MS/MS and provides an
-application programming interface (API) for users to define their own
-spectrum preprocessing transformations and similarity measures
-[@Huber2020]. The pyOpenMS tool, which is a Python wrapper of the OpenMS
-tool, is for analyzing LC-MS/MS data and provides common spectrum
-preprocessing transformations such as filtering based on m/z and
-intensity values [@Rost2014; @Rost2016]. The Python package
-spectrum_utils provides functionality for preprocessing and visualizing
-mass spectrometry data [@Bittremieux2020]. The R package metID is a tool
-for LC-MS/MS-based compound identification with a focus on allowing users
-to combine their in-house database(s) with public databases [@Shen2022].
-The R Shiny package ShinyMetID provides users six similarity measures
-(Cosine, Weighted Cosine, Stein & Scott, Discrete Fourier
-Transformation, Discrete Wavelet Transformation, and Semi-Partial
-Correlation) and a graphical user interface (GUI) to perform spectral
-library matching to identify chemical compounds from GC-MS data
-[@Jeong2023].
[…]
