+++ paper/paper.md
@@ -47,7 +47,7 @@ To address the lack of spectral library matching software that considers both th
 
 **Table  1.** Comparison of MZsearch and existing tools.
 
-![Comparison of MZsearch and existing tools.\label{fig:comparison}](table.PNG){width="110%," height="110%"}
+![\label{fig:comparison}](table.PNG){width="100%," height="100%"}
 
 # Functionality
 
