+++ paper/paper.md
@@ -45,7 +45,7 @@ Several open-source software tools have been developed for analyzing mass spectr
 In many GC-MS experiments, the resulting mass spectra often have m/z values reported as integers due to the use of nominal mass resolution of mass spectrometers. In contrast, LC-MS/MS frequently employs a high-resolution mass spectrometer that provides m/z values with multiple decimal places. This necessitates several spectrum preprocessing steps unique to the high-resolution MS or MS/MS data analysis, namely centroiding (i.e., merging peaks that are 'close' with respect to their m/z values) and matching (i.e., aligning the m/z values so that the query spectrum and reference spectrum have the same length). In addition to these canonical spectrum preprocessing transformations for high-resolution MS/MS data, weight factor transformations and low-entropy transformations have been proposed to improve the performance of compound identification for both GC-MS and LC-MS/MS data [@Kim2012; @Li2021; @Dlugas2024_preprint]. The Shannon Entropy Similarity Measure has been shown to outperform the Cosine Similarity Measure with respect to LC-MS/MS data [@Li2021]. A generalization of the Shannon Entropy Similarity Measure, the Tsallis Entropy Similarity Measure, slightly outperforms the Shannon Entropy Similarity Measure in both GC-MS and LC-MS/MS data [@Dlugas2024_preprint]. This recent study has further demonstrated that the order of preprocessing transformations is also critical to achieving higher accuracy in compound identification. However, to our knowledge, there is no tool that includes the recently-developed, high-performance entropy-based similarity measures or considers the order of the spectrum preprocessing transformations, which are important for effective metabolomics studies.
 To address the lack of spectral library matching software that considers both the order of spectrum preprocessing steps and novel entropy-based similarity measures, MZsearch was developed. The developed MZsearch is a Python-based tool for performing spectral library matching on both nominal-resolution mass spectrometry (NR-MS) data (e.g. GC-MS) and high-resolution mass spectrometry  (HR-MS) data (e.g., LC-MS/MS). It is available in two versions: a command-line interface and Python modules for integration into custom code. Users can customize their own spectrum preprocessing order using spectrum preprocessing transformations such as weight factor and low-entropy transformations. Additionally, MZsearch includes a novel entropy-based similarity measure, the Rényi Entropy Similarity Measure, enabling users to choose among four similarity measures: the commonly-used Cosine Similarity Measure [@Stein1994], the Shannon Entropy Similarity Measure [@Li2021], the recently-developed Tsallis Entropy Similarity Measure [@Dlugas2024_preprint], and the novel Rényi Entropy Similarity Measure. **Table 1** compares MZsearch with other similar software.
 
-**Table  1.** Comparison of MZsearch and existing tools.
+<p align="center"> **Table  1.** Comparison of MZsearch and existing tools. </p>
 
 ![\label{comparison}](table.PNG){width="100%," height="100%"}\
 
