### 2025-04-03 #7594 @hdlugas
@jromanowska

We, the authors, would like to thank the Editor and the two reviewers for their thorough review and constructive suggestions on both the software package and the manuscript. We also appreciate the Editor’s help in converting our previous responses to the GitHub thread.

In the first round of responses, we used a single file that combined all comments and responses, as this made communication among the three authors more convenient. For the same reason, we felt that attaching a single consolidated file would also make it easier for the Editor and reviewers to review our responses. We apologize for not realizing at the time that this approach was inconsistent with the open-source review process.

In this second round, we followed a similar approach during our internal discussion. After coordinating among all three authors, we have addressed each specific point raised, and all collected comments along with our corresponding replies are included below.

@lionel42
**Q1:** My comment Q2 was not understood. There are 3 main differences between GC-MS vs LC-MS, low res (integer masses) vs highres (float masses) and MS1 vs MS2 (also called tandem mass spectromety, see https://en.wikipedia.org/wiki/Tandem_mass_spectrometry ). I have seen that many changes in the code have modified the lcms to lcmsms making us understand that MS2 is used with LC, and MS1 with GC. However as I tried to explain, the chromatographic (also called separation) part and the fragmentation part (high res vs low res and MS1 vs MS2) are independant . The current text (starting around line 30) in the paper is completely confusing and seems to have been written by someone that is very confused about the science behind. Also the data for LCMS provided does not correspond to the MS2 data type as the new lcmsms notation suggests

**Response:** The reviewer is correct. The resolution of a mass spectrum is determined by the mass spectrometer. To date, all LC-MS instruments are equipped with a high-resolution mass spectrometer. The majority of GC-MS systems are equipped with mass spectrometers with a nominal mass resolution. In GC-MS, electron ionization (EI) with a collision energy of 70 eV is widely employed. The EI can directly fragment the molecular ion to generate the fragment ion spectrum. We highly appreciate the reviewer’s comment and revised the manuscript accordingly. Additionally, in the code, we use NRMS to represent the nominal-resolution mass spectra and HRMS for high-resolution mass spectra. We also emphasize that MZsearch offers a range of spectrum preprocessing transformations and similarity measures for analysis of data generated by a mass spectrometer with nominal mass resolution and a high-resolution mass spectrometer.

**Q2:** In the updated table, I see now that matchms is not compatible with GC-MS, which is wrong

**Response:** We originally indicated that matchms was not compatible with GC-MS based on the first sentence of the README file in matchms’s GitHub repository which states: “Matchms is a versatile open-source Python package developed for importing, processing, cleaning, and comparing mass spectrometry data (MS/MS)”. Since it explicitly refers to MS/MS and not MS, we assumed the package was compatible with higher-resolution MS data and not nominal MS data. Based on the reviewer’s comment, we have revised the table and now indicate that matchms is compatible with HR-MS as well as NR-MS.

**Q3:** The updated table contains the same confusion with GC-MS vs LC-MS instead of MS vs MS2 as I suggested in the previous comment

**Response:** As noted in our response to the first comment in this round, we have now revised the terminology throughout the code and manuscript to distinguish based on resolution and fragmentation level, using HR-MS (or HRMS in the code) to refer to high-resolution mass spectra and NR-MS (or NRMS in the code) for nominal-resolution mass spectra.

**Q4:** The python interface usage proposed is very wired (https://github.com/hdlugas/MZsearch?tab=readme-ov-file#322-python-interface). It does not follow any python good practice. (* character in the import, path to the code must be specified instead of installable package)

**Response:** Considering that MZsearch is intended to primarily be a command line tool, we provide the example python script only to illustrate how a user may use MZsearch’s functionality in their custom python code. Care was taken to ensure that function names in the two imported python modules are all dissimilar to avoid naming conflicts.

**Q5:** I do not understand why [@hdlugas](https://github.com/hdlugas) does not want to use github to provide answers, answers which sound very "AI generated".

**Response:** In the first round of responses, we used a single file that combined all comments and replies, as this made communication among the three authors more convenient. For the same reason, we felt that attaching a single consolidated file would also make it easier for the Editor and reviewers to follow our responses. We apologize for not realizing at the time that this approach was inconsistent with the open-source review process. We also want to clarify that all responses were addressed by the authors and were not generated by AI.

**Q6:** The updated table is visually aweful which seems as if no one took care of having looked at it.

**Response:** We have now created the table as an image.

**Q7:** Duplicate code in spec_lib_matching.py for the transformations which appeared to be very similar for HRMS and LRMS. Also in test_similarity_measures.py

**Response:** Due to (i) the slightly different data format for high-resolution MS (HR-MS) data and nominal-resolution MS (NR-MS) data and (ii) more preprocessing transformations available for HR-MS, the cases of HR-MS and NR-MS are considered separately. The script ‘tests_similarity_measures.py’ provides two examples as a sanity check to ensure that the similarity measures behave as expected (i.e. similar spectra have a similarity close to 1 and dissimilar spectra have a similarity close to 0).

**Q8:** Poor exception handling when bad values are given see details (Handling of bad parameters hdlugas/MZsearch#2)

**Response:** Many checks have now been incorporated to give specific errors related to exception handling.

**Q9:** Poor usage of the argparse module. https://docs.python.org/3/library/argparse.html The argparse module lets developper add many custom info on variables such as types or default values. In this package (which claims that one of the main improvement to other packages is the CLI) only a very basic usage is done (the help).

**Response:** The functionality of the argparse module which allows the developer to specify the type and default value of a CLI-input parameter has been included where possible. Additionally, more tests which experiment with exception handling are included as well.

**Q10:** Usage of inline comments instead of docstings, as commonly done in python

**Response:** Comments spanning multiple lines have been converted from comments with a pound sign to docstrings.

@aboffin

**Q1:** Removing the description from the table (as a response to question 1), was never requested by the other reviewer, as far as I understand. In fact, a clear and accurate description helps the readers in understanding the application of the various tools listed in the table.

**Response:** We followed the Editor’s suggestion not to include the descriptions in the table. While we agree that the description can help readers better understand the applications of the listed tools, we believe that the current table still provides sufficient information for readers to grasp the general characteristics and purposes of each tool in the context of compound identification. Therefore, we have decided to adhere to the Editor’s guidance on this point.

**Q2:** In Question 5, I did not mention absolute path, I asked the authors to fix the relative paths so that the code can be run from the repo.

**Response:** The examples should now run with the current paths.

**Q3:** Have all the authors reviewed the manuscript and consent to the content and changes made by the first author? Usually a statement: "All authors have reviewed the final version of the manuscript." or something similar is included in papers, not sure about JOSS's editorial policy on this.

**Response:** In fact, we included the statement “H.D., X.Z., and S.K. reviewed and edited the manuscript” in the Author Contributions section of the previous version of the manuscript, which conveys the same meaning.

**Q4:** •	Questions about AI have been raised by both [@lionel42](https://github.com/lionel42) and me. In their response, they clarify that AI was not used to generate the code, however, this is the second or third time a partial,"almost quite, but not entirely unlike" list of responses to reviewers' comments were provided that causes issues. Perhaps this avoidance of addressing all the comments is not intentional but it still is a cause for concern.

**Response:** As we stated before, all responses were addressed by the authors and not by AI. We apologize if we misunderstood the comments, resulting in partially addressed responses. In this second round, we have paid much more attention to each reviewer’s comment and tried to address them fully.
