### 2025-04-27 #7594 @Kevin-Mattheus-Moerman
@hdlugas the editorial board has studied and re-evaluated this submission, it's size, functionality offered. In addition we've studied this review and the issues raised. The board has concluded that unfortunately this submission does not pass our [substantial scholarly effort criteria](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort). As such we will now proceed to reject this submission.
